Enrich an rlang traceback with details on native calls

This function uses the native stack trace returned from winch_trace_back() to add details on native function calls to an rlang traceback object. It is intended to be called by rlang.

winch_add_trace_back(trace = rlang::trace_back(bottom = parent.frame()))

Arguments

trace
An rlang traceback as returned by rlang::trace_back().

Examples

foo <- function() {
  winch_call(bar)
}

bar <- function() {
  trace <- rlang::trace_back()
  winch_add_trace_back(trace)
}

foo()
#>      █
#>   1. └─pkgdown::deploy_to_branch(new_process = FALSE)
#>   2.   └─pkgdown::build_site(...)
#>   3.     └─pkgdown:::build_site_local(...)
#>   4.       └─pkgdown::build_reference(...)
#>   5.         └─purrr::map(...)
#>   6.           └─`/purrr.so`::map_impl()
#>   7.             └─`/purrr.so`::call_loop()
#>   8.               └─pkgdown:::.f(.x[[i]], ...)
#>   9.                 ├─base::withCallingHandlers(...)
#>  10.                 └─pkgdown:::data_reference_topic(...)
#>  11.                   └─pkgdown:::run_examples(...)
#>  12.                     └─pkgdown:::highlight_examples(code, topic, env = env)
#>  13.                       └─downlit::evaluate_and_highlight(...)
#>  14.                         └─evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#>  15.                           └─evaluate:::evaluate_call(...)
#>  16.                             ├─evaluate:::timing_fn(...)
#>  17.                             ├─evaluate:::handle(...)
#>  18.                             │ └─base::try(f, silent = TRUE)
#>  19.                             │   └─base::tryCatch(...)
#>  20.                             │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  21.                             │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  22.                             │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  23.                             ├─base::withCallingHandlers(...)
#>  24.                             ├─base::withVisible(eval(expr, envir, enclos))
#>  25.                             └─base::eval(expr, envir, enclos)
#>  26.                               └─base::eval(expr, envir, enclos)
#>  27.                                 └─foo()
#>  28.                                   └─winch::winch_call(bar)
#>  29.                                     └─`/winch.so`::winch_call()
#>  30.                                       └─(function () ... /home/runner/work/winch/winch/R/call.R:23:2