Native stack trace

This function returns the native stack trace as a data frame. Each native stack frame corresponds to one row in the returned data frame. Deep function calls come first, the last row corresponds to the running process's entry point.

winch_trace_back()

Value

A data frame with the columns:

func: function name

ip: instruction pointer

pathname: path to shared library

Details

On Windows, call winch_init_library() to return function names for a specific package.

See also

Examples

winch_trace_back()

#>                  func               ip
#> 1              bcEval 00007f8368b75260
#> 2             Rf_eval 00007f8368b8abd0
#> 3       R_execClosure 00007f8368b8c850
#> 4     Rf_applyClosure 00007f8368b8d770
#> 5             Rf_eval 00007f8368b8abd0
#> 6             do_eval 00007f8368b90a80
#> 7              bcEval 00007f8368b75260
#> 8             Rf_eval 00007f8368b8abd0
#> 9       R_execClosure 00007f8368b8c850
#> 10    Rf_applyClosure 00007f8368b8d770
#> 11             bcEval 00007f8368b75260
#> 12            Rf_eval 00007f8368b8abd0
#> 13       forcePromise 00007f8368b8b640
#> 14            Rf_eval 00007f8368b8abd0
#> 15     do_withVisible 00007f8368b91850
#> 16        do_internal 00007f8368bce6e0
#> 17             bcEval 00007f8368b75260
#> 18            Rf_eval 00007f8368b8abd0
#> 19      R_execClosure 00007f8368b8c850
#> 20    Rf_applyClosure 00007f8368b8d770
#> 21             bcEval 00007f8368b75260
#> 22            Rf_eval 00007f8368b8abd0
#> 23       forcePromise 00007f8368b8b640
#> 24             getvar 00007f8368b8b780
#> 25             bcEval 00007f8368b75260
#> 26            Rf_eval 00007f8368b8abd0
#> 27      R_execClosure 00007f8368b8c850
#> 28    Rf_applyClosure 00007f8368b8d770
#> 29             bcEval 00007f8368b75260
#> 30            Rf_eval 00007f8368b8abd0
#> 31       forcePromise 00007f8368b8b640
#> 32             getvar 00007f8368b8b780
#> 33             bcEval 00007f8368b75260
#> 34            Rf_eval 00007f8368b8abd0
#> 35       forcePromise 00007f8368b8b640
#> 36             getvar 00007f8368b8b780
#> 37             bcEval 00007f8368b75260
#> 38            Rf_eval 00007f8368b8abd0
#> 39       forcePromise 00007f8368b8b640
#> 40             getvar 00007f8368b8b780
#> 41             bcEval 00007f8368b75260
#> 42            Rf_eval 00007f8368b8abd0
#> 43       forcePromise 00007f8368b8b640
#> 44             getvar 00007f8368b8b780
#> 45             bcEval 00007f8368b75260
#> 46            Rf_eval 00007f8368b8abd0
#> 47       forcePromise 00007f8368b8b640
#> 48             getvar 00007f8368b8b780
#> 49             bcEval 00007f8368b75260
#> 50            Rf_eval 00007f8368b8abd0
#> 51       forcePromise 00007f8368b8b640
#> 52             getvar 00007f8368b8b780
#> 53             bcEval 00007f8368b75260
#> 54            Rf_eval 00007f8368b8abd0
#> 55      R_execClosure 00007f8368b8c850
#> 56    Rf_applyClosure 00007f8368b8d770
#> 57             bcEval 00007f8368b75260
#> 58            Rf_eval 00007f8368b8abd0
#> 59      R_execClosure 00007f8368b8c850
#> 60    Rf_applyClosure 00007f8368b8d770
#> 61             bcEval 00007f8368b75260
#> 62            Rf_eval 00007f8368b8abd0
#> 63      R_execClosure 00007f8368b8c850
#> 64    Rf_applyClosure 00007f8368b8d770
#> 65             bcEval 00007f8368b75260
#> 66            Rf_eval 00007f8368b8abd0
#> 67      R_execClosure 00007f8368b8c850
#> 68    Rf_applyClosure 00007f8368b8d770
#> 69             bcEval 00007f8368b75260
#> 70            Rf_eval 00007f8368b8abd0
#> 71      R_execClosure 00007f8368b8c850
#> 72    Rf_applyClosure 00007f8368b8d770
#> 73             bcEval 00007f8368b75260
#> 74            Rf_eval 00007f8368b8abd0
#> 75      R_execClosure 00007f8368b8c850
#> 76    Rf_applyClosure 00007f8368b8d770
#> 77             bcEval 00007f8368b75260
#> 78            Rf_eval 00007f8368b8abd0
#> 79       forcePromise 00007f8368b8b640
#> 80             getvar 00007f8368b8b780
#> 81             bcEval 00007f8368b75260
#> 82            Rf_eval 00007f8368b8abd0
#> 83      R_execClosure 00007f8368b8c850
#> 84    Rf_applyClosure 00007f8368b8d770
#> 85             bcEval 00007f8368b75260
#> 86            Rf_eval 00007f8368b8abd0
#> 87      R_execClosure 00007f8368b8c850
#> 88    Rf_applyClosure 00007f8368b8d770
#> 89             bcEval 00007f8368b75260
#> 90            Rf_eval 00007f8368b8abd0
#> 91      R_execClosure 00007f8368b8c850
#> 92    Rf_applyClosure 00007f8368b8d770
#> 93             bcEval 00007f8368b75260
#> 94            Rf_eval 00007f8368b8abd0
#> 95      R_execClosure 00007f8368b8c850
#> 96    Rf_applyClosure 00007f8368b8d770
#> 97             bcEval 00007f8368b75260
#> 98            Rf_eval 00007f8368b8abd0
#> 99      R_execClosure 00007f8368b8c850
#> 100   Rf_applyClosure 00007f8368b8d770
#> 101            bcEval 00007f8368b75260
#> 102           Rf_eval 00007f8368b8abd0
#> 103     R_execClosure 00007f8368b8c850
#> 104   Rf_applyClosure 00007f8368b8d770
#> 105            bcEval 00007f8368b75260
#> 106           Rf_eval 00007f8368b8abd0
#> 107     R_execClosure 00007f8368b8c850
#> 108   Rf_applyClosure 00007f8368b8d770
#> 109            bcEval 00007f8368b75260
#> 110           Rf_eval 00007f8368b8abd0
#> 111      forcePromise 00007f8368b8b640
#> 112            getvar 00007f8368b8b780
#> 113            bcEval 00007f8368b75260
#> 114           Rf_eval 00007f8368b8abd0
#> 115     R_execClosure 00007f8368b8c850
#> 116   Rf_applyClosure 00007f8368b8d770
#> 117            bcEval 00007f8368b75260
#> 118           Rf_eval 00007f8368b8abd0
#> 119     R_execClosure 00007f8368b8c850
#> 120   Rf_applyClosure 00007f8368b8d770
#> 121    R_forceAndCall 00007f8368b90140
#> 122         call_loop 00007f8356c813b0
#> 123          map_impl 00007f8356c81500
#> 124       R_doDotCall 00007f8368b3ef80
#> 125            bcEval 00007f8368b75260
#> 126           Rf_eval 00007f8368b8abd0
#> 127     R_execClosure 00007f8368b8c850
#> 128   Rf_applyClosure 00007f8368b8d770
#> 129            bcEval 00007f8368b75260
#> 130           Rf_eval 00007f8368b8abd0
#> 131     R_execClosure 00007f8368b8c850
#> 132   Rf_applyClosure 00007f8368b8d770
#> 133            bcEval 00007f8368b75260
#> 134           Rf_eval 00007f8368b8abd0
#> 135     R_execClosure 00007f8368b8c850
#> 136   Rf_applyClosure 00007f8368b8d770
#> 137            bcEval 00007f8368b75260
#> 138           Rf_eval 00007f8368b8abd0
#> 139     R_execClosure 00007f8368b8c850
#> 140   Rf_applyClosure 00007f8368b8d770
#> 141            bcEval 00007f8368b75260
#> 142           Rf_eval 00007f8368b8abd0
#> 143     R_execClosure 00007f8368b8c850
#> 144   Rf_applyClosure 00007f8368b8d770
#> 145           Rf_eval 00007f8368b8abd0
#> 146  Rf_ReplIteration 00007f8368bbe840
#> 147     R_ReplConsole 00007f8368bbedc0
#> 148     run_Rmainloop 00007f8368bbeeb0
#> 149              main 00005579c2d1f820
#> 150 __libc_start_main 00007f8368671ab0
#> 151            _start 00005579c2d1f850
#>                                                pathname
#> 1                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 2                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 3                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 4                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 5                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 6                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 7                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 8                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 9                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 10                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 11                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 12                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 13                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 14                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 15                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 16                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 17                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 18                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 19                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 20                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 21                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 22                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 23                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 24                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 25                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 26                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 27                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 28                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 29                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 30                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 31                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 32                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 33                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 34                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 35                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 36                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 37                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 38                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 39                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 40                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 41                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 42                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 43                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 44                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 45                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 46                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 47                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 48                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 49                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 50                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 51                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 52                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 53                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 54                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 55                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 56                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 57                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 58                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 59                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 60                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 61                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 62                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 63                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 64                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 65                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 66                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 67                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 68                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 69                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 70                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 71                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 72                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 73                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 74                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 75                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 76                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 77                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 78                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 79                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 80                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 81                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 82                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 83                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 84                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 85                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 86                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 87                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 88                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 89                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 90                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 91                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 92                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 93                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 94                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 95                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 96                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 97                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 98                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 99                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 100                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 101                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 102                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 103                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 104                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 105                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 106                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 107                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 108                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 109                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 110                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 111                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 112                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 113                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 114                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 115                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 116                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 117                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 118                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 119                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 120                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 121                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 122 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 123 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 124                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 125                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 126                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 127                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 128                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 129                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 130                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 131                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 132                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 133                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 134                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 135                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 136                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 137                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 138                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 139                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 140                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 141                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 142                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 143                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 144                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 145                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 146                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 147                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 148                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 149                       /opt/R/4.0.2/lib/R/bin/exec/R
#> 150                  /lib/x86_64-linux-gnu/libc-2.27.so
#> 151                       /opt/R/4.0.2/lib/R/bin/exec/R

foo <- function() {
  winch_call(bar)
}

bar <- function() {
  winch_trace_back()
}

foo()

#>                  func               ip
#> 1              bcEval 00007f8368b75260
#> 2             Rf_eval 00007f8368b8abd0
#> 3       R_execClosure 00007f8368b8c850
#> 4     Rf_applyClosure 00007f8368b8d770
#> 5             Rf_eval 00007f8368b8abd0
#> 6            do_begin 00007f8368b8e570
#> 7             Rf_eval 00007f8368b8abd0
#> 8       R_execClosure 00007f8368b8c850
#> 9     Rf_applyClosure 00007f8368b8d770
#> 10            Rf_eval 00007f8368b8abd0
#> 11         winch_call 00007f835016c5b0
#> 12        R_doDotCall 00007f8368b3ef80
#> 13             bcEval 00007f8368b75260
#> 14            Rf_eval 00007f8368b8abd0
#> 15      R_execClosure 00007f8368b8c850
#> 16    Rf_applyClosure 00007f8368b8d770
#> 17            Rf_eval 00007f8368b8abd0
#> 18           do_begin 00007f8368b8e570
#> 19            Rf_eval 00007f8368b8abd0
#> 20      R_execClosure 00007f8368b8c850
#> 21    Rf_applyClosure 00007f8368b8d770
#> 22            Rf_eval 00007f8368b8abd0
#> 23            do_eval 00007f8368b90a80
#> 24             bcEval 00007f8368b75260
#> 25            Rf_eval 00007f8368b8abd0
#> 26      R_execClosure 00007f8368b8c850
#> 27    Rf_applyClosure 00007f8368b8d770
#> 28             bcEval 00007f8368b75260
#> 29            Rf_eval 00007f8368b8abd0
#> 30       forcePromise 00007f8368b8b640
#> 31            Rf_eval 00007f8368b8abd0
#> 32     do_withVisible 00007f8368b91850
#> 33        do_internal 00007f8368bce6e0
#> 34             bcEval 00007f8368b75260
#> 35            Rf_eval 00007f8368b8abd0
#> 36      R_execClosure 00007f8368b8c850
#> 37    Rf_applyClosure 00007f8368b8d770
#> 38             bcEval 00007f8368b75260
#> 39            Rf_eval 00007f8368b8abd0
#> 40       forcePromise 00007f8368b8b640
#> 41             getvar 00007f8368b8b780
#> 42             bcEval 00007f8368b75260
#> 43            Rf_eval 00007f8368b8abd0
#> 44      R_execClosure 00007f8368b8c850
#> 45    Rf_applyClosure 00007f8368b8d770
#> 46             bcEval 00007f8368b75260
#> 47            Rf_eval 00007f8368b8abd0
#> 48       forcePromise 00007f8368b8b640
#> 49             getvar 00007f8368b8b780
#> 50             bcEval 00007f8368b75260
#> 51            Rf_eval 00007f8368b8abd0
#> 52       forcePromise 00007f8368b8b640
#> 53             getvar 00007f8368b8b780
#> 54             bcEval 00007f8368b75260
#> 55            Rf_eval 00007f8368b8abd0
#> 56       forcePromise 00007f8368b8b640
#> 57             getvar 00007f8368b8b780
#> 58             bcEval 00007f8368b75260
#> 59            Rf_eval 00007f8368b8abd0
#> 60       forcePromise 00007f8368b8b640
#> 61             getvar 00007f8368b8b780
#> 62             bcEval 00007f8368b75260
#> 63            Rf_eval 00007f8368b8abd0
#> 64       forcePromise 00007f8368b8b640
#> 65             getvar 00007f8368b8b780
#> 66             bcEval 00007f8368b75260
#> 67            Rf_eval 00007f8368b8abd0
#> 68       forcePromise 00007f8368b8b640
#> 69             getvar 00007f8368b8b780
#> 70             bcEval 00007f8368b75260
#> 71            Rf_eval 00007f8368b8abd0
#> 72      R_execClosure 00007f8368b8c850
#> 73    Rf_applyClosure 00007f8368b8d770
#> 74             bcEval 00007f8368b75260
#> 75            Rf_eval 00007f8368b8abd0
#> 76      R_execClosure 00007f8368b8c850
#> 77    Rf_applyClosure 00007f8368b8d770
#> 78             bcEval 00007f8368b75260
#> 79            Rf_eval 00007f8368b8abd0
#> 80      R_execClosure 00007f8368b8c850
#> 81    Rf_applyClosure 00007f8368b8d770
#> 82             bcEval 00007f8368b75260
#> 83            Rf_eval 00007f8368b8abd0
#> 84      R_execClosure 00007f8368b8c850
#> 85    Rf_applyClosure 00007f8368b8d770
#> 86             bcEval 00007f8368b75260
#> 87            Rf_eval 00007f8368b8abd0
#> 88      R_execClosure 00007f8368b8c850
#> 89    Rf_applyClosure 00007f8368b8d770
#> 90             bcEval 00007f8368b75260
#> 91            Rf_eval 00007f8368b8abd0
#> 92      R_execClosure 00007f8368b8c850
#> 93    Rf_applyClosure 00007f8368b8d770
#> 94             bcEval 00007f8368b75260
#> 95            Rf_eval 00007f8368b8abd0
#> 96       forcePromise 00007f8368b8b640
#> 97             getvar 00007f8368b8b780
#> 98             bcEval 00007f8368b75260
#> 99            Rf_eval 00007f8368b8abd0
#> 100     R_execClosure 00007f8368b8c850
#> 101   Rf_applyClosure 00007f8368b8d770
#> 102            bcEval 00007f8368b75260
#> 103           Rf_eval 00007f8368b8abd0
#> 104     R_execClosure 00007f8368b8c850
#> 105   Rf_applyClosure 00007f8368b8d770
#> 106            bcEval 00007f8368b75260
#> 107           Rf_eval 00007f8368b8abd0
#> 108     R_execClosure 00007f8368b8c850
#> 109   Rf_applyClosure 00007f8368b8d770
#> 110            bcEval 00007f8368b75260
#> 111           Rf_eval 00007f8368b8abd0
#> 112     R_execClosure 00007f8368b8c850
#> 113   Rf_applyClosure 00007f8368b8d770
#> 114            bcEval 00007f8368b75260
#> 115           Rf_eval 00007f8368b8abd0
#> 116     R_execClosure 00007f8368b8c850
#> 117   Rf_applyClosure 00007f8368b8d770
#> 118            bcEval 00007f8368b75260
#> 119           Rf_eval 00007f8368b8abd0
#> 120     R_execClosure 00007f8368b8c850
#> 121   Rf_applyClosure 00007f8368b8d770
#> 122            bcEval 00007f8368b75260
#> 123           Rf_eval 00007f8368b8abd0
#> 124     R_execClosure 00007f8368b8c850
#> 125   Rf_applyClosure 00007f8368b8d770
#> 126            bcEval 00007f8368b75260
#> 127           Rf_eval 00007f8368b8abd0
#> 128      forcePromise 00007f8368b8b640
#> 129            getvar 00007f8368b8b780
#> 130            bcEval 00007f8368b75260
#> 131           Rf_eval 00007f8368b8abd0
#> 132     R_execClosure 00007f8368b8c850
#> 133   Rf_applyClosure 00007f8368b8d770
#> 134            bcEval 00007f8368b75260
#> 135           Rf_eval 00007f8368b8abd0
#> 136     R_execClosure 00007f8368b8c850
#> 137   Rf_applyClosure 00007f8368b8d770
#> 138    R_forceAndCall 00007f8368b90140
#> 139         call_loop 00007f8356c813b0
#> 140          map_impl 00007f8356c81500
#> 141       R_doDotCall 00007f8368b3ef80
#> 142            bcEval 00007f8368b75260
#> 143           Rf_eval 00007f8368b8abd0
#> 144     R_execClosure 00007f8368b8c850
#> 145   Rf_applyClosure 00007f8368b8d770
#> 146            bcEval 00007f8368b75260
#> 147           Rf_eval 00007f8368b8abd0
#> 148     R_execClosure 00007f8368b8c850
#> 149   Rf_applyClosure 00007f8368b8d770
#> 150            bcEval 00007f8368b75260
#> 151           Rf_eval 00007f8368b8abd0
#> 152     R_execClosure 00007f8368b8c850
#> 153   Rf_applyClosure 00007f8368b8d770
#> 154            bcEval 00007f8368b75260
#> 155           Rf_eval 00007f8368b8abd0
#> 156     R_execClosure 00007f8368b8c850
#> 157   Rf_applyClosure 00007f8368b8d770
#> 158            bcEval 00007f8368b75260
#> 159           Rf_eval 00007f8368b8abd0
#> 160     R_execClosure 00007f8368b8c850
#> 161   Rf_applyClosure 00007f8368b8d770
#> 162           Rf_eval 00007f8368b8abd0
#> 163  Rf_ReplIteration 00007f8368bbe840
#> 164     R_ReplConsole 00007f8368bbedc0
#> 165     run_Rmainloop 00007f8368bbeeb0
#> 166              main 00005579c2d1f820
#> 167 __libc_start_main 00007f8368671ab0
#> 168            _start 00005579c2d1f850
#>                                                pathname
#> 1                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 2                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 3                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 4                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 5                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 6                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 7                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 8                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 9                        /opt/R/4.0.2/lib/R/lib/libR.so
#> 10                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 11  /home/runner/work/_temp/Library/winch/libs/winch.so
#> 12                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 13                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 14                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 15                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 16                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 17                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 18                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 19                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 20                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 21                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 22                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 23                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 24                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 25                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 26                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 27                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 28                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 29                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 30                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 31                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 32                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 33                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 34                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 35                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 36                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 37                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 38                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 39                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 40                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 41                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 42                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 43                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 44                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 45                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 46                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 47                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 48                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 49                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 50                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 51                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 52                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 53                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 54                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 55                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 56                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 57                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 58                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 59                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 60                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 61                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 62                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 63                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 64                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 65                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 66                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 67                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 68                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 69                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 70                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 71                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 72                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 73                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 74                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 75                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 76                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 77                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 78                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 79                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 80                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 81                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 82                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 83                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 84                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 85                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 86                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 87                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 88                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 89                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 90                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 91                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 92                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 93                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 94                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 95                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 96                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 97                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 98                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 99                       /opt/R/4.0.2/lib/R/lib/libR.so
#> 100                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 101                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 102                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 103                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 104                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 105                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 106                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 107                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 108                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 109                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 110                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 111                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 112                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 113                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 114                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 115                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 116                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 117                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 118                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 119                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 120                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 121                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 122                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 123                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 124                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 125                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 126                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 127                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 128                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 129                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 130                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 131                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 132                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 133                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 134                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 135                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 136                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 137                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 138                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 139 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 140 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 141                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 142                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 143                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 144                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 145                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 146                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 147                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 148                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 149                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 150                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 151                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 152                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 153                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 154                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 155                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 156                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 157                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 158                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 159                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 160                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 161                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 162                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 163                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 164                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 165                      /opt/R/4.0.2/lib/R/lib/libR.so
#> 166                       /opt/R/4.0.2/lib/R/bin/exec/R
#> 167                  /lib/x86_64-linux-gnu/libc-2.27.so
#> 168                       /opt/R/4.0.2/lib/R/bin/exec/R