Enrich an rlang traceback with details on native calls

This function uses the native stack trace returned from winch_trace_back() to add details on native function calls to an rlang traceback object. It is intended to be called by rlang.

winch_add_trace_back(trace = rlang::trace_back(bottom = parent.frame()))

Arguments

trace

An rlang traceback as returned by rlang::trace_back().

Examples

foo <- function() {
  winch_call(bar)
}

bar <- function() {
  trace <- rlang::trace_back()
  winch_add_trace_back(trace)
}

foo()
#>      ▆
#>   1. └─pkgdown::deploy_to_branch(new_process = FALSE)
#>   2.   └─pkgdown::build_site_github_pages(...)
#>   3.     └─pkgdown::build_site(...)
#>   4.       └─pkgdown:::build_site_local(...)
#>   5.         └─pkgdown::build_reference(...)
#>   6.           └─purrr::map(...)
#>   7.             └─`/purrr.so`::map_impl()
#>   8.               └─`/purrr.so`::call_loop()
#>   9.                 └─pkgdown .f(.x[[i]], ...)
#>  10.                   ├─base::withCallingHandlers(...)
#>  11.                   └─pkgdown:::data_reference_topic(...)
#>  12.                     └─pkgdown:::run_examples(...)
#>  13.                       └─pkgdown:::highlight_examples(code, topic, env = env)
#>  14.                         └─downlit::evaluate_and_highlight(...)
#>  15.                           └─evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#>  16.                             └─evaluate:::evaluate_call(...)
#>  17.                               ├─evaluate timing_fn(...)
#>  18.                               ├─evaluate handle(...)
#>  19.                               │ └─base::try(f, silent = TRUE)
#>  20.                               │   └─base::tryCatch(...)
#>  21.                               │     └─base tryCatchList(expr, classes, parentenv, handlers)
#>  22.                               │       └─base tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  23.                               │         └─base doTryCatch(return(expr), name, parentenv, handler)
#>  24.                               ├─base::withCallingHandlers(...)
#>  25.                               ├─base::withVisible(eval(expr, envir, enclos))
#>  26.                               └─base::eval(expr, envir, enclos)
#>  27.                                 └─base::eval(expr, envir, enclos)
#>  28.                                   └─foo()
#>  29.                                     └─winch::winch_call(bar)
#>  30.                                       └─`/winch.so`::winch_call()
#>  31.                                         └─`<fn>`() at winch/R/call.R:23:2