Native stack trace

This function returns the native stack trace as a data frame. Each native stack frame corresponds to one row in the returned data frame. Deep function calls come first, the last row corresponds to the running process's entry point.

winch_trace_back()

Value

A data frame with the columns:

func: function name

ip: instruction pointer

pathname: path to shared library

is_libr: a logical, TRUE if this entry is from R's shared library, determined via procmaps::path_is_libr() on the pathname component

Details

On Windows, call winch_init_library() to return function names for a specific package.

See also

Examples

winch_trace_back()
#>                                 func               ip
#> 1                      R_ParseString 00007f2972f333d0
#> 2                            Rf_eval 00007f2972f52b80
#> 3                            Rf_eval 00007f2972f54af0
#> 4                    Rf_applyClosure 00007f2972f55a20
#> 5                            Rf_eval 00007f2972f52b80
#> 6                     R_forceAndCall 00007f2972f58560
#> 7                      R_ParseString 00007f2972f333d0
#> 8                            Rf_eval 00007f2972f52b80
#> 9                            Rf_eval 00007f2972f54af0
#> 10                   Rf_applyClosure 00007f2972f55a20
#> 11                     R_ParseString 00007f2972f333d0
#> 12                           Rf_eval 00007f2972f52b80
#> 13                           Rf_eval 00007f2972f54af0
#> 14                   Rf_applyClosure 00007f2972f55a20
#> 15                     R_ParseString 00007f2972f333d0
#> 16                           Rf_eval 00007f2972f52b80
#> 17                           Rf_eval 00007f2972f535f0
#> 18                           Rf_eval 00007f2972f52b80
#> 19                    R_forceAndCall 00007f2972f59060
#> 20                       do_Rprofmem 00007f2972f98c40
#> 21                     R_ParseString 00007f2972f333d0
#> 22                           Rf_eval 00007f2972f52b80
#> 23                           Rf_eval 00007f2972f54af0
#> 24                   Rf_applyClosure 00007f2972f55a20
#> 25                     R_ParseString 00007f2972f333d0
#> 26                           Rf_eval 00007f2972f52b80
#> 27                           Rf_eval 00007f2972f535f0
#> 28                           Rf_eval 00007f2972f53720
#> 29                     R_ParseString 00007f2972f333d0
#> 30                           Rf_eval 00007f2972f52b80
#> 31                           Rf_eval 00007f2972f54af0
#> 32                   Rf_applyClosure 00007f2972f55a20
#> 33                     R_ParseString 00007f2972f333d0
#> 34                           Rf_eval 00007f2972f52b80
#> 35                           Rf_eval 00007f2972f535f0
#> 36                           Rf_eval 00007f2972f53720
#> 37                     R_ParseString 00007f2972f333d0
#> 38                           Rf_eval 00007f2972f52b80
#> 39                           Rf_eval 00007f2972f535f0
#> 40                           Rf_eval 00007f2972f53720
#> 41                     R_ParseString 00007f2972f333d0
#> 42                           Rf_eval 00007f2972f52b80
#> 43                           Rf_eval 00007f2972f535f0
#> 44                           Rf_eval 00007f2972f53720
#> 45                     R_ParseString 00007f2972f333d0
#> 46                           Rf_eval 00007f2972f52b80
#> 47                           Rf_eval 00007f2972f535f0
#> 48                           Rf_eval 00007f2972f53720
#> 49                     R_ParseString 00007f2972f333d0
#> 50                           Rf_eval 00007f2972f52b80
#> 51                           Rf_eval 00007f2972f535f0
#> 52                           Rf_eval 00007f2972f53720
#> 53                     R_ParseString 00007f2972f333d0
#> 54                           Rf_eval 00007f2972f52b80
#> 55                           Rf_eval 00007f2972f535f0
#> 56                           Rf_eval 00007f2972f53720
#> 57                     R_ParseString 00007f2972f333d0
#> 58                           Rf_eval 00007f2972f52b80
#> 59                           Rf_eval 00007f2972f54af0
#> 60                   Rf_applyClosure 00007f2972f55a20
#> 61                     R_ParseString 00007f2972f333d0
#> 62                           Rf_eval 00007f2972f52b80
#> 63                           Rf_eval 00007f2972f54af0
#> 64                   Rf_applyClosure 00007f2972f55a20
#> 65                     R_ParseString 00007f2972f333d0
#> 66                           Rf_eval 00007f2972f52b80
#> 67                           Rf_eval 00007f2972f54af0
#> 68                   Rf_applyClosure 00007f2972f55a20
#> 69                     R_ParseString 00007f2972f333d0
#> 70                           Rf_eval 00007f2972f52b80
#> 71                           Rf_eval 00007f2972f54af0
#> 72                   Rf_applyClosure 00007f2972f55a20
#> 73                     R_ParseString 00007f2972f333d0
#> 74                           Rf_eval 00007f2972f52b80
#> 75                           Rf_eval 00007f2972f54af0
#> 76                   Rf_applyClosure 00007f2972f55a20
#> 77                     R_ParseString 00007f2972f333d0
#> 78                           Rf_eval 00007f2972f52b80
#> 79                           Rf_eval 00007f2972f54af0
#> 80                   Rf_applyClosure 00007f2972f55a20
#> 81                     R_ParseString 00007f2972f333d0
#> 82                           Rf_eval 00007f2972f52b80
#> 83                           Rf_eval 00007f2972f535f0
#> 84                           Rf_eval 00007f2972f53720
#> 85                     R_ParseString 00007f2972f333d0
#> 86                           Rf_eval 00007f2972f52b80
#> 87                           Rf_eval 00007f2972f54af0
#> 88                   Rf_applyClosure 00007f2972f55a20
#> 89                     R_ParseString 00007f2972f333d0
#> 90                           Rf_eval 00007f2972f52b80
#> 91                           Rf_eval 00007f2972f54af0
#> 92                   Rf_applyClosure 00007f2972f55a20
#> 93                     R_ParseString 00007f2972f333d0
#> 94                           Rf_eval 00007f2972f52b80
#> 95                           Rf_eval 00007f2972f54af0
#> 96                   Rf_applyClosure 00007f2972f55a20
#> 97                     R_ParseString 00007f2972f333d0
#> 98                           Rf_eval 00007f2972f52b80
#> 99                           Rf_eval 00007f2972f54af0
#> 100                  Rf_applyClosure 00007f2972f55a20
#> 101                    R_ParseString 00007f2972f333d0
#> 102                          Rf_eval 00007f2972f52b80
#> 103                          Rf_eval 00007f2972f54af0
#> 104                  Rf_applyClosure 00007f2972f55a20
#> 105                    R_ParseString 00007f2972f333d0
#> 106                          Rf_eval 00007f2972f52b80
#> 107                          Rf_eval 00007f2972f54af0
#> 108                  Rf_applyClosure 00007f2972f55a20
#> 109                    R_ParseString 00007f2972f333d0
#> 110                          Rf_eval 00007f2972f52b80
#> 111                          Rf_eval 00007f2972f54af0
#> 112                  Rf_applyClosure 00007f2972f55a20
#> 113                    R_ParseString 00007f2972f333d0
#> 114                          Rf_eval 00007f2972f52b80
#> 115                          Rf_eval 00007f2972f535f0
#> 116                          Rf_eval 00007f2972f53720
#> 117                    R_ParseString 00007f2972f333d0
#> 118                          Rf_eval 00007f2972f52b80
#> 119                          Rf_eval 00007f2972f54af0
#> 120                  Rf_applyClosure 00007f2972f55a20
#> 121                    R_ParseString 00007f2972f333d0
#> 122                          Rf_eval 00007f2972f52b80
#> 123                          Rf_eval 00007f2972f54af0
#> 124                  Rf_applyClosure 00007f2972f55a20
#> 125                   R_forceAndCall 00007f2972f57cd0
#> 126                     R_init_purrr 00007f29700eebf0
#> 127                     R_init_purrr 00007f29700eee00
#> 128               Rf_NewFrameConfirm 00007f2972ef9920
#> 129               Rf_NewFrameConfirm 00007f2972efb4a0
#> 130                          Rf_eval 00007f2972f52b80
#> 131 <unw_get_proc_name() error: -10> 00007f29700ecc20
#> 132                R_ExecWithCleanup 00007f2972ee25c0
#> 133 <unw_get_proc_name() error: -10> 00007f29700eceb0
#> 134 <unw_get_proc_name() error: -10> 00007f29700ecf60
#> 135               Rf_NewFrameConfirm 00007f2972ef9920
#> 136                    R_ParseString 00007f2972f333d0
#> 137                          Rf_eval 00007f2972f52b80
#> 138                          Rf_eval 00007f2972f54af0
#> 139                  Rf_applyClosure 00007f2972f55a20
#> 140                    R_ParseString 00007f2972f333d0
#> 141                          Rf_eval 00007f2972f52b80
#> 142                          Rf_eval 00007f2972f535f0
#> 143                          Rf_eval 00007f2972f53720
#> 144                    R_ParseString 00007f2972f333d0
#> 145                          Rf_eval 00007f2972f52b80
#> 146                          Rf_eval 00007f2972f535f0
#> 147                          Rf_eval 00007f2972f53720
#> 148                    R_ParseString 00007f2972f333d0
#> 149                          Rf_eval 00007f2972f52b80
#> 150                          Rf_eval 00007f2972f54af0
#> 151                  Rf_applyClosure 00007f2972f55a20
#> 152                    R_ParseString 00007f2972f333d0
#> 153                          Rf_eval 00007f2972f52b80
#> 154                          Rf_eval 00007f2972f54af0
#> 155                  Rf_applyClosure 00007f2972f55a20
#> 156                    R_ParseString 00007f2972f333d0
#> 157                          Rf_eval 00007f2972f52b80
#> 158                          Rf_eval 00007f2972f54af0
#> 159                  Rf_applyClosure 00007f2972f55a20
#> 160                    R_ParseString 00007f2972f333d0
#> 161                          Rf_eval 00007f2972f52b80
#> 162                          Rf_eval 00007f2972f54af0
#> 163                  Rf_applyClosure 00007f2972f55a20
#> 164                    R_ParseString 00007f2972f333d0
#> 165                          Rf_eval 00007f2972f52b80
#> 166                          Rf_eval 00007f2972f54af0
#> 167                  Rf_applyClosure 00007f2972f55a20
#> 168                    R_ParseString 00007f2972f333d0
#> 169                          Rf_eval 00007f2972f52b80
#> 170                          Rf_eval 00007f2972f54af0
#> 171                  Rf_applyClosure 00007f2972f55a20
#> 172                    R_ParseString 00007f2972f333d0
#> 173                          Rf_eval 00007f2972f52b80
#> 174                          Rf_eval 00007f2972f54af0
#> 175                  Rf_applyClosure 00007f2972f55a20
#> 176                    R_ParseString 00007f2972f333d0
#> 177                          Rf_eval 00007f2972f52b80
#> 178                          Rf_eval 00007f2972f54af0
#> 179                  Rf_applyClosure 00007f2972f55a20
#> 180                    R_ParseString 00007f2972f333d0
#> 181                          Rf_eval 00007f2972f52b80
#> 182                          Rf_eval 00007f2972f54af0
#> 183                  Rf_applyClosure 00007f2972f55a20
#> 184                          Rf_eval 00007f2972f52b80
#> 185                 Rf_ReplIteration 00007f2972f88b10
#> 186                 Rf_ReplIteration 00007f2972f89020
#> 187                    run_Rmainloop 00007f2972f89120
#> 188                             main 000055e3bac02080
#> 189                __libc_init_first 00007f2972a29d10
#> 190                __libc_start_main 00007f2972a29dc0
#> 191                           _start 000055e3bac020b0
#>                                                pathname is_libr
#> 1                                /usr/lib/R/lib/libR.so    TRUE
#> 2                                /usr/lib/R/lib/libR.so    TRUE
#> 3                                /usr/lib/R/lib/libR.so    TRUE
#> 4                                /usr/lib/R/lib/libR.so    TRUE
#> 5                                /usr/lib/R/lib/libR.so    TRUE
#> 6                                /usr/lib/R/lib/libR.so    TRUE
#> 7                                /usr/lib/R/lib/libR.so    TRUE
#> 8                                /usr/lib/R/lib/libR.so    TRUE
#> 9                                /usr/lib/R/lib/libR.so    TRUE
#> 10                               /usr/lib/R/lib/libR.so    TRUE
#> 11                               /usr/lib/R/lib/libR.so    TRUE
#> 12                               /usr/lib/R/lib/libR.so    TRUE
#> 13                               /usr/lib/R/lib/libR.so    TRUE
#> 14                               /usr/lib/R/lib/libR.so    TRUE
#> 15                               /usr/lib/R/lib/libR.so    TRUE
#> 16                               /usr/lib/R/lib/libR.so    TRUE
#> 17                               /usr/lib/R/lib/libR.so    TRUE
#> 18                               /usr/lib/R/lib/libR.so    TRUE
#> 19                               /usr/lib/R/lib/libR.so    TRUE
#> 20                               /usr/lib/R/lib/libR.so    TRUE
#> 21                               /usr/lib/R/lib/libR.so    TRUE
#> 22                               /usr/lib/R/lib/libR.so    TRUE
#> 23                               /usr/lib/R/lib/libR.so    TRUE
#> 24                               /usr/lib/R/lib/libR.so    TRUE
#> 25                               /usr/lib/R/lib/libR.so    TRUE
#> 26                               /usr/lib/R/lib/libR.so    TRUE
#> 27                               /usr/lib/R/lib/libR.so    TRUE
#> 28                               /usr/lib/R/lib/libR.so    TRUE
#> 29                               /usr/lib/R/lib/libR.so    TRUE
#> 30                               /usr/lib/R/lib/libR.so    TRUE
#> 31                               /usr/lib/R/lib/libR.so    TRUE
#> 32                               /usr/lib/R/lib/libR.so    TRUE
#> 33                               /usr/lib/R/lib/libR.so    TRUE
#> 34                               /usr/lib/R/lib/libR.so    TRUE
#> 35                               /usr/lib/R/lib/libR.so    TRUE
#> 36                               /usr/lib/R/lib/libR.so    TRUE
#> 37                               /usr/lib/R/lib/libR.so    TRUE
#> 38                               /usr/lib/R/lib/libR.so    TRUE
#> 39                               /usr/lib/R/lib/libR.so    TRUE
#> 40                               /usr/lib/R/lib/libR.so    TRUE
#> 41                               /usr/lib/R/lib/libR.so    TRUE
#> 42                               /usr/lib/R/lib/libR.so    TRUE
#> 43                               /usr/lib/R/lib/libR.so    TRUE
#> 44                               /usr/lib/R/lib/libR.so    TRUE
#> 45                               /usr/lib/R/lib/libR.so    TRUE
#> 46                               /usr/lib/R/lib/libR.so    TRUE
#> 47                               /usr/lib/R/lib/libR.so    TRUE
#> 48                               /usr/lib/R/lib/libR.so    TRUE
#> 49                               /usr/lib/R/lib/libR.so    TRUE
#> 50                               /usr/lib/R/lib/libR.so    TRUE
#> 51                               /usr/lib/R/lib/libR.so    TRUE
#> 52                               /usr/lib/R/lib/libR.so    TRUE
#> 53                               /usr/lib/R/lib/libR.so    TRUE
#> 54                               /usr/lib/R/lib/libR.so    TRUE
#> 55                               /usr/lib/R/lib/libR.so    TRUE
#> 56                               /usr/lib/R/lib/libR.so    TRUE
#> 57                               /usr/lib/R/lib/libR.so    TRUE
#> 58                               /usr/lib/R/lib/libR.so    TRUE
#> 59                               /usr/lib/R/lib/libR.so    TRUE
#> 60                               /usr/lib/R/lib/libR.so    TRUE
#> 61                               /usr/lib/R/lib/libR.so    TRUE
#> 62                               /usr/lib/R/lib/libR.so    TRUE
#> 63                               /usr/lib/R/lib/libR.so    TRUE
#> 64                               /usr/lib/R/lib/libR.so    TRUE
#> 65                               /usr/lib/R/lib/libR.so    TRUE
#> 66                               /usr/lib/R/lib/libR.so    TRUE
#> 67                               /usr/lib/R/lib/libR.so    TRUE
#> 68                               /usr/lib/R/lib/libR.so    TRUE
#> 69                               /usr/lib/R/lib/libR.so    TRUE
#> 70                               /usr/lib/R/lib/libR.so    TRUE
#> 71                               /usr/lib/R/lib/libR.so    TRUE
#> 72                               /usr/lib/R/lib/libR.so    TRUE
#> 73                               /usr/lib/R/lib/libR.so    TRUE
#> 74                               /usr/lib/R/lib/libR.so    TRUE
#> 75                               /usr/lib/R/lib/libR.so    TRUE
#> 76                               /usr/lib/R/lib/libR.so    TRUE
#> 77                               /usr/lib/R/lib/libR.so    TRUE
#> 78                               /usr/lib/R/lib/libR.so    TRUE
#> 79                               /usr/lib/R/lib/libR.so    TRUE
#> 80                               /usr/lib/R/lib/libR.so    TRUE
#> 81                               /usr/lib/R/lib/libR.so    TRUE
#> 82                               /usr/lib/R/lib/libR.so    TRUE
#> 83                               /usr/lib/R/lib/libR.so    TRUE
#> 84                               /usr/lib/R/lib/libR.so    TRUE
#> 85                               /usr/lib/R/lib/libR.so    TRUE
#> 86                               /usr/lib/R/lib/libR.so    TRUE
#> 87                               /usr/lib/R/lib/libR.so    TRUE
#> 88                               /usr/lib/R/lib/libR.so    TRUE
#> 89                               /usr/lib/R/lib/libR.so    TRUE
#> 90                               /usr/lib/R/lib/libR.so    TRUE
#> 91                               /usr/lib/R/lib/libR.so    TRUE
#> 92                               /usr/lib/R/lib/libR.so    TRUE
#> 93                               /usr/lib/R/lib/libR.so    TRUE
#> 94                               /usr/lib/R/lib/libR.so    TRUE
#> 95                               /usr/lib/R/lib/libR.so    TRUE
#> 96                               /usr/lib/R/lib/libR.so    TRUE
#> 97                               /usr/lib/R/lib/libR.so    TRUE
#> 98                               /usr/lib/R/lib/libR.so    TRUE
#> 99                               /usr/lib/R/lib/libR.so    TRUE
#> 100                              /usr/lib/R/lib/libR.so    TRUE
#> 101                              /usr/lib/R/lib/libR.so    TRUE
#> 102                              /usr/lib/R/lib/libR.so    TRUE
#> 103                              /usr/lib/R/lib/libR.so    TRUE
#> 104                              /usr/lib/R/lib/libR.so    TRUE
#> 105                              /usr/lib/R/lib/libR.so    TRUE
#> 106                              /usr/lib/R/lib/libR.so    TRUE
#> 107                              /usr/lib/R/lib/libR.so    TRUE
#> 108                              /usr/lib/R/lib/libR.so    TRUE
#> 109                              /usr/lib/R/lib/libR.so    TRUE
#> 110                              /usr/lib/R/lib/libR.so    TRUE
#> 111                              /usr/lib/R/lib/libR.so    TRUE
#> 112                              /usr/lib/R/lib/libR.so    TRUE
#> 113                              /usr/lib/R/lib/libR.so    TRUE
#> 114                              /usr/lib/R/lib/libR.so    TRUE
#> 115                              /usr/lib/R/lib/libR.so    TRUE
#> 116                              /usr/lib/R/lib/libR.so    TRUE
#> 117                              /usr/lib/R/lib/libR.so    TRUE
#> 118                              /usr/lib/R/lib/libR.so    TRUE
#> 119                              /usr/lib/R/lib/libR.so    TRUE
#> 120                              /usr/lib/R/lib/libR.so    TRUE
#> 121                              /usr/lib/R/lib/libR.so    TRUE
#> 122                              /usr/lib/R/lib/libR.so    TRUE
#> 123                              /usr/lib/R/lib/libR.so    TRUE
#> 124                              /usr/lib/R/lib/libR.so    TRUE
#> 125                              /usr/lib/R/lib/libR.so    TRUE
#> 126 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 127 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 128                              /usr/lib/R/lib/libR.so    TRUE
#> 129                              /usr/lib/R/lib/libR.so    TRUE
#> 130                              /usr/lib/R/lib/libR.so    TRUE
#> 131 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 132                              /usr/lib/R/lib/libR.so    TRUE
#> 133 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 134 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 135                              /usr/lib/R/lib/libR.so    TRUE
#> 136                              /usr/lib/R/lib/libR.so    TRUE
#> 137                              /usr/lib/R/lib/libR.so    TRUE
#> 138                              /usr/lib/R/lib/libR.so    TRUE
#> 139                              /usr/lib/R/lib/libR.so    TRUE
#> 140                              /usr/lib/R/lib/libR.so    TRUE
#> 141                              /usr/lib/R/lib/libR.so    TRUE
#> 142                              /usr/lib/R/lib/libR.so    TRUE
#> 143                              /usr/lib/R/lib/libR.so    TRUE
#> 144                              /usr/lib/R/lib/libR.so    TRUE
#> 145                              /usr/lib/R/lib/libR.so    TRUE
#> 146                              /usr/lib/R/lib/libR.so    TRUE
#> 147                              /usr/lib/R/lib/libR.so    TRUE
#> 148                              /usr/lib/R/lib/libR.so    TRUE
#> 149                              /usr/lib/R/lib/libR.so    TRUE
#> 150                              /usr/lib/R/lib/libR.so    TRUE
#> 151                              /usr/lib/R/lib/libR.so    TRUE
#> 152                              /usr/lib/R/lib/libR.so    TRUE
#> 153                              /usr/lib/R/lib/libR.so    TRUE
#> 154                              /usr/lib/R/lib/libR.so    TRUE
#> 155                              /usr/lib/R/lib/libR.so    TRUE
#> 156                              /usr/lib/R/lib/libR.so    TRUE
#> 157                              /usr/lib/R/lib/libR.so    TRUE
#> 158                              /usr/lib/R/lib/libR.so    TRUE
#> 159                              /usr/lib/R/lib/libR.so    TRUE
#> 160                              /usr/lib/R/lib/libR.so    TRUE
#> 161                              /usr/lib/R/lib/libR.so    TRUE
#> 162                              /usr/lib/R/lib/libR.so    TRUE
#> 163                              /usr/lib/R/lib/libR.so    TRUE
#> 164                              /usr/lib/R/lib/libR.so    TRUE
#> 165                              /usr/lib/R/lib/libR.so    TRUE
#> 166                              /usr/lib/R/lib/libR.so    TRUE
#> 167                              /usr/lib/R/lib/libR.so    TRUE
#> 168                              /usr/lib/R/lib/libR.so    TRUE
#> 169                              /usr/lib/R/lib/libR.so    TRUE
#> 170                              /usr/lib/R/lib/libR.so    TRUE
#> 171                              /usr/lib/R/lib/libR.so    TRUE
#> 172                              /usr/lib/R/lib/libR.so    TRUE
#> 173                              /usr/lib/R/lib/libR.so    TRUE
#> 174                              /usr/lib/R/lib/libR.so    TRUE
#> 175                              /usr/lib/R/lib/libR.so    TRUE
#> 176                              /usr/lib/R/lib/libR.so    TRUE
#> 177                              /usr/lib/R/lib/libR.so    TRUE
#> 178                              /usr/lib/R/lib/libR.so    TRUE
#> 179                              /usr/lib/R/lib/libR.so    TRUE
#> 180                              /usr/lib/R/lib/libR.so    TRUE
#> 181                              /usr/lib/R/lib/libR.so    TRUE
#> 182                              /usr/lib/R/lib/libR.so    TRUE
#> 183                              /usr/lib/R/lib/libR.so    TRUE
#> 184                              /usr/lib/R/lib/libR.so    TRUE
#> 185                              /usr/lib/R/lib/libR.so    TRUE
#> 186                              /usr/lib/R/lib/libR.so    TRUE
#> 187                              /usr/lib/R/lib/libR.so    TRUE
#> 188                               /usr/lib/R/bin/exec/R    TRUE
#> 189                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 190                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 191                               /usr/lib/R/bin/exec/R    TRUE

foo <- function() {
  winch_call(bar)
}

bar <- function() {
  winch_trace_back()
}

foo()
#>                                 func               ip
#> 1                      R_ParseString 00007f2972f333d0
#> 2                            Rf_eval 00007f2972f52b80
#> 3                            Rf_eval 00007f2972f54af0
#> 4                    Rf_applyClosure 00007f2972f55a20
#> 5                            Rf_eval 00007f2972f52b80
#> 6                       R_execMethod 00007f2972f562b0
#> 7                            Rf_eval 00007f2972f52b80
#> 8                            Rf_eval 00007f2972f54af0
#> 9                    Rf_applyClosure 00007f2972f55a20
#> 10                           Rf_eval 00007f2972f52b80
#> 11                        winch_call 00007f29678b7710
#> 12                Rf_NewFrameConfirm 00007f2972ef9920
#> 13                     R_ParseString 00007f2972f333d0
#> 14                           Rf_eval 00007f2972f52b80
#> 15                           Rf_eval 00007f2972f54af0
#> 16                   Rf_applyClosure 00007f2972f55a20
#> 17                           Rf_eval 00007f2972f52b80
#> 18                      R_execMethod 00007f2972f562b0
#> 19                           Rf_eval 00007f2972f52b80
#> 20                           Rf_eval 00007f2972f54af0
#> 21                   Rf_applyClosure 00007f2972f55a20
#> 22                           Rf_eval 00007f2972f52b80
#> 23                    R_forceAndCall 00007f2972f58560
#> 24                     R_ParseString 00007f2972f333d0
#> 25                           Rf_eval 00007f2972f52b80
#> 26                           Rf_eval 00007f2972f54af0
#> 27                   Rf_applyClosure 00007f2972f55a20
#> 28                     R_ParseString 00007f2972f333d0
#> 29                           Rf_eval 00007f2972f52b80
#> 30                           Rf_eval 00007f2972f54af0
#> 31                   Rf_applyClosure 00007f2972f55a20
#> 32                     R_ParseString 00007f2972f333d0
#> 33                           Rf_eval 00007f2972f52b80
#> 34                           Rf_eval 00007f2972f535f0
#> 35                           Rf_eval 00007f2972f52b80
#> 36                    R_forceAndCall 00007f2972f59060
#> 37                       do_Rprofmem 00007f2972f98c40
#> 38                     R_ParseString 00007f2972f333d0
#> 39                           Rf_eval 00007f2972f52b80
#> 40                           Rf_eval 00007f2972f54af0
#> 41                   Rf_applyClosure 00007f2972f55a20
#> 42                     R_ParseString 00007f2972f333d0
#> 43                           Rf_eval 00007f2972f52b80
#> 44                           Rf_eval 00007f2972f535f0
#> 45                           Rf_eval 00007f2972f53720
#> 46                     R_ParseString 00007f2972f333d0
#> 47                           Rf_eval 00007f2972f52b80
#> 48                           Rf_eval 00007f2972f54af0
#> 49                   Rf_applyClosure 00007f2972f55a20
#> 50                     R_ParseString 00007f2972f333d0
#> 51                           Rf_eval 00007f2972f52b80
#> 52                           Rf_eval 00007f2972f535f0
#> 53                           Rf_eval 00007f2972f53720
#> 54                     R_ParseString 00007f2972f333d0
#> 55                           Rf_eval 00007f2972f52b80
#> 56                           Rf_eval 00007f2972f535f0
#> 57                           Rf_eval 00007f2972f53720
#> 58                     R_ParseString 00007f2972f333d0
#> 59                           Rf_eval 00007f2972f52b80
#> 60                           Rf_eval 00007f2972f535f0
#> 61                           Rf_eval 00007f2972f53720
#> 62                     R_ParseString 00007f2972f333d0
#> 63                           Rf_eval 00007f2972f52b80
#> 64                           Rf_eval 00007f2972f535f0
#> 65                           Rf_eval 00007f2972f53720
#> 66                     R_ParseString 00007f2972f333d0
#> 67                           Rf_eval 00007f2972f52b80
#> 68                           Rf_eval 00007f2972f535f0
#> 69                           Rf_eval 00007f2972f53720
#> 70                     R_ParseString 00007f2972f333d0
#> 71                           Rf_eval 00007f2972f52b80
#> 72                           Rf_eval 00007f2972f535f0
#> 73                           Rf_eval 00007f2972f53720
#> 74                     R_ParseString 00007f2972f333d0
#> 75                           Rf_eval 00007f2972f52b80
#> 76                           Rf_eval 00007f2972f54af0
#> 77                   Rf_applyClosure 00007f2972f55a20
#> 78                     R_ParseString 00007f2972f333d0
#> 79                           Rf_eval 00007f2972f52b80
#> 80                           Rf_eval 00007f2972f54af0
#> 81                   Rf_applyClosure 00007f2972f55a20
#> 82                     R_ParseString 00007f2972f333d0
#> 83                           Rf_eval 00007f2972f52b80
#> 84                           Rf_eval 00007f2972f54af0
#> 85                   Rf_applyClosure 00007f2972f55a20
#> 86                     R_ParseString 00007f2972f333d0
#> 87                           Rf_eval 00007f2972f52b80
#> 88                           Rf_eval 00007f2972f54af0
#> 89                   Rf_applyClosure 00007f2972f55a20
#> 90                     R_ParseString 00007f2972f333d0
#> 91                           Rf_eval 00007f2972f52b80
#> 92                           Rf_eval 00007f2972f54af0
#> 93                   Rf_applyClosure 00007f2972f55a20
#> 94                     R_ParseString 00007f2972f333d0
#> 95                           Rf_eval 00007f2972f52b80
#> 96                           Rf_eval 00007f2972f54af0
#> 97                   Rf_applyClosure 00007f2972f55a20
#> 98                     R_ParseString 00007f2972f333d0
#> 99                           Rf_eval 00007f2972f52b80
#> 100                          Rf_eval 00007f2972f535f0
#> 101                          Rf_eval 00007f2972f53720
#> 102                    R_ParseString 00007f2972f333d0
#> 103                          Rf_eval 00007f2972f52b80
#> 104                          Rf_eval 00007f2972f54af0
#> 105                  Rf_applyClosure 00007f2972f55a20
#> 106                    R_ParseString 00007f2972f333d0
#> 107                          Rf_eval 00007f2972f52b80
#> 108                          Rf_eval 00007f2972f54af0
#> 109                  Rf_applyClosure 00007f2972f55a20
#> 110                    R_ParseString 00007f2972f333d0
#> 111                          Rf_eval 00007f2972f52b80
#> 112                          Rf_eval 00007f2972f54af0
#> 113                  Rf_applyClosure 00007f2972f55a20
#> 114                    R_ParseString 00007f2972f333d0
#> 115                          Rf_eval 00007f2972f52b80
#> 116                          Rf_eval 00007f2972f54af0
#> 117                  Rf_applyClosure 00007f2972f55a20
#> 118                    R_ParseString 00007f2972f333d0
#> 119                          Rf_eval 00007f2972f52b80
#> 120                          Rf_eval 00007f2972f54af0
#> 121                  Rf_applyClosure 00007f2972f55a20
#> 122                    R_ParseString 00007f2972f333d0
#> 123                          Rf_eval 00007f2972f52b80
#> 124                          Rf_eval 00007f2972f54af0
#> 125                  Rf_applyClosure 00007f2972f55a20
#> 126                    R_ParseString 00007f2972f333d0
#> 127                          Rf_eval 00007f2972f52b80
#> 128                          Rf_eval 00007f2972f54af0
#> 129                  Rf_applyClosure 00007f2972f55a20
#> 130                    R_ParseString 00007f2972f333d0
#> 131                          Rf_eval 00007f2972f52b80
#> 132                          Rf_eval 00007f2972f535f0
#> 133                          Rf_eval 00007f2972f53720
#> 134                    R_ParseString 00007f2972f333d0
#> 135                          Rf_eval 00007f2972f52b80
#> 136                          Rf_eval 00007f2972f54af0
#> 137                  Rf_applyClosure 00007f2972f55a20
#> 138                    R_ParseString 00007f2972f333d0
#> 139                          Rf_eval 00007f2972f52b80
#> 140                          Rf_eval 00007f2972f54af0
#> 141                  Rf_applyClosure 00007f2972f55a20
#> 142                   R_forceAndCall 00007f2972f57cd0
#> 143                     R_init_purrr 00007f29700eebf0
#> 144                     R_init_purrr 00007f29700eee00
#> 145               Rf_NewFrameConfirm 00007f2972ef9920
#> 146               Rf_NewFrameConfirm 00007f2972efb4a0
#> 147                          Rf_eval 00007f2972f52b80
#> 148 <unw_get_proc_name() error: -10> 00007f29700ecc20
#> 149                R_ExecWithCleanup 00007f2972ee25c0
#> 150 <unw_get_proc_name() error: -10> 00007f29700eceb0
#> 151 <unw_get_proc_name() error: -10> 00007f29700ecf60
#> 152               Rf_NewFrameConfirm 00007f2972ef9920
#> 153                    R_ParseString 00007f2972f333d0
#> 154                          Rf_eval 00007f2972f52b80
#> 155                          Rf_eval 00007f2972f54af0
#> 156                  Rf_applyClosure 00007f2972f55a20
#> 157                    R_ParseString 00007f2972f333d0
#> 158                          Rf_eval 00007f2972f52b80
#> 159                          Rf_eval 00007f2972f535f0
#> 160                          Rf_eval 00007f2972f53720
#> 161                    R_ParseString 00007f2972f333d0
#> 162                          Rf_eval 00007f2972f52b80
#> 163                          Rf_eval 00007f2972f535f0
#> 164                          Rf_eval 00007f2972f53720
#> 165                    R_ParseString 00007f2972f333d0
#> 166                          Rf_eval 00007f2972f52b80
#> 167                          Rf_eval 00007f2972f54af0
#> 168                  Rf_applyClosure 00007f2972f55a20
#> 169                    R_ParseString 00007f2972f333d0
#> 170                          Rf_eval 00007f2972f52b80
#> 171                          Rf_eval 00007f2972f54af0
#> 172                  Rf_applyClosure 00007f2972f55a20
#> 173                    R_ParseString 00007f2972f333d0
#> 174                          Rf_eval 00007f2972f52b80
#> 175                          Rf_eval 00007f2972f54af0
#> 176                  Rf_applyClosure 00007f2972f55a20
#> 177                    R_ParseString 00007f2972f333d0
#> 178                          Rf_eval 00007f2972f52b80
#> 179                          Rf_eval 00007f2972f54af0
#> 180                  Rf_applyClosure 00007f2972f55a20
#> 181                    R_ParseString 00007f2972f333d0
#> 182                          Rf_eval 00007f2972f52b80
#> 183                          Rf_eval 00007f2972f54af0
#> 184                  Rf_applyClosure 00007f2972f55a20
#> 185                    R_ParseString 00007f2972f333d0
#> 186                          Rf_eval 00007f2972f52b80
#> 187                          Rf_eval 00007f2972f54af0
#> 188                  Rf_applyClosure 00007f2972f55a20
#> 189                    R_ParseString 00007f2972f333d0
#> 190                          Rf_eval 00007f2972f52b80
#> 191                          Rf_eval 00007f2972f54af0
#> 192                  Rf_applyClosure 00007f2972f55a20
#> 193                    R_ParseString 00007f2972f333d0
#> 194                          Rf_eval 00007f2972f52b80
#> 195                          Rf_eval 00007f2972f54af0
#> 196                  Rf_applyClosure 00007f2972f55a20
#> 197                    R_ParseString 00007f2972f333d0
#> 198                          Rf_eval 00007f2972f52b80
#> 199                          Rf_eval 00007f2972f54af0
#> 200                  Rf_applyClosure 00007f2972f55a20
#> 201                          Rf_eval 00007f2972f52b80
#> 202                 Rf_ReplIteration 00007f2972f88b10
#> 203                 Rf_ReplIteration 00007f2972f89020
#> 204                    run_Rmainloop 00007f2972f89120
#> 205                             main 000055e3bac02080
#> 206                __libc_init_first 00007f2972a29d10
#> 207                __libc_start_main 00007f2972a29dc0
#> 208                           _start 000055e3bac020b0
#>                                                pathname is_libr
#> 1                                /usr/lib/R/lib/libR.so    TRUE
#> 2                                /usr/lib/R/lib/libR.so    TRUE
#> 3                                /usr/lib/R/lib/libR.so    TRUE
#> 4                                /usr/lib/R/lib/libR.so    TRUE
#> 5                                /usr/lib/R/lib/libR.so    TRUE
#> 6                                /usr/lib/R/lib/libR.so    TRUE
#> 7                                /usr/lib/R/lib/libR.so    TRUE
#> 8                                /usr/lib/R/lib/libR.so    TRUE
#> 9                                /usr/lib/R/lib/libR.so    TRUE
#> 10                               /usr/lib/R/lib/libR.so    TRUE
#> 11  /home/runner/work/_temp/Library/winch/libs/winch.so   FALSE
#> 12                               /usr/lib/R/lib/libR.so    TRUE
#> 13                               /usr/lib/R/lib/libR.so    TRUE
#> 14                               /usr/lib/R/lib/libR.so    TRUE
#> 15                               /usr/lib/R/lib/libR.so    TRUE
#> 16                               /usr/lib/R/lib/libR.so    TRUE
#> 17                               /usr/lib/R/lib/libR.so    TRUE
#> 18                               /usr/lib/R/lib/libR.so    TRUE
#> 19                               /usr/lib/R/lib/libR.so    TRUE
#> 20                               /usr/lib/R/lib/libR.so    TRUE
#> 21                               /usr/lib/R/lib/libR.so    TRUE
#> 22                               /usr/lib/R/lib/libR.so    TRUE
#> 23                               /usr/lib/R/lib/libR.so    TRUE
#> 24                               /usr/lib/R/lib/libR.so    TRUE
#> 25                               /usr/lib/R/lib/libR.so    TRUE
#> 26                               /usr/lib/R/lib/libR.so    TRUE
#> 27                               /usr/lib/R/lib/libR.so    TRUE
#> 28                               /usr/lib/R/lib/libR.so    TRUE
#> 29                               /usr/lib/R/lib/libR.so    TRUE
#> 30                               /usr/lib/R/lib/libR.so    TRUE
#> 31                               /usr/lib/R/lib/libR.so    TRUE
#> 32                               /usr/lib/R/lib/libR.so    TRUE
#> 33                               /usr/lib/R/lib/libR.so    TRUE
#> 34                               /usr/lib/R/lib/libR.so    TRUE
#> 35                               /usr/lib/R/lib/libR.so    TRUE
#> 36                               /usr/lib/R/lib/libR.so    TRUE
#> 37                               /usr/lib/R/lib/libR.so    TRUE
#> 38                               /usr/lib/R/lib/libR.so    TRUE
#> 39                               /usr/lib/R/lib/libR.so    TRUE
#> 40                               /usr/lib/R/lib/libR.so    TRUE
#> 41                               /usr/lib/R/lib/libR.so    TRUE
#> 42                               /usr/lib/R/lib/libR.so    TRUE
#> 43                               /usr/lib/R/lib/libR.so    TRUE
#> 44                               /usr/lib/R/lib/libR.so    TRUE
#> 45                               /usr/lib/R/lib/libR.so    TRUE
#> 46                               /usr/lib/R/lib/libR.so    TRUE
#> 47                               /usr/lib/R/lib/libR.so    TRUE
#> 48                               /usr/lib/R/lib/libR.so    TRUE
#> 49                               /usr/lib/R/lib/libR.so    TRUE
#> 50                               /usr/lib/R/lib/libR.so    TRUE
#> 51                               /usr/lib/R/lib/libR.so    TRUE
#> 52                               /usr/lib/R/lib/libR.so    TRUE
#> 53                               /usr/lib/R/lib/libR.so    TRUE
#> 54                               /usr/lib/R/lib/libR.so    TRUE
#> 55                               /usr/lib/R/lib/libR.so    TRUE
#> 56                               /usr/lib/R/lib/libR.so    TRUE
#> 57                               /usr/lib/R/lib/libR.so    TRUE
#> 58                               /usr/lib/R/lib/libR.so    TRUE
#> 59                               /usr/lib/R/lib/libR.so    TRUE
#> 60                               /usr/lib/R/lib/libR.so    TRUE
#> 61                               /usr/lib/R/lib/libR.so    TRUE
#> 62                               /usr/lib/R/lib/libR.so    TRUE
#> 63                               /usr/lib/R/lib/libR.so    TRUE
#> 64                               /usr/lib/R/lib/libR.so    TRUE
#> 65                               /usr/lib/R/lib/libR.so    TRUE
#> 66                               /usr/lib/R/lib/libR.so    TRUE
#> 67                               /usr/lib/R/lib/libR.so    TRUE
#> 68                               /usr/lib/R/lib/libR.so    TRUE
#> 69                               /usr/lib/R/lib/libR.so    TRUE
#> 70                               /usr/lib/R/lib/libR.so    TRUE
#> 71                               /usr/lib/R/lib/libR.so    TRUE
#> 72                               /usr/lib/R/lib/libR.so    TRUE
#> 73                               /usr/lib/R/lib/libR.so    TRUE
#> 74                               /usr/lib/R/lib/libR.so    TRUE
#> 75                               /usr/lib/R/lib/libR.so    TRUE
#> 76                               /usr/lib/R/lib/libR.so    TRUE
#> 77                               /usr/lib/R/lib/libR.so    TRUE
#> 78                               /usr/lib/R/lib/libR.so    TRUE
#> 79                               /usr/lib/R/lib/libR.so    TRUE
#> 80                               /usr/lib/R/lib/libR.so    TRUE
#> 81                               /usr/lib/R/lib/libR.so    TRUE
#> 82                               /usr/lib/R/lib/libR.so    TRUE
#> 83                               /usr/lib/R/lib/libR.so    TRUE
#> 84                               /usr/lib/R/lib/libR.so    TRUE
#> 85                               /usr/lib/R/lib/libR.so    TRUE
#> 86                               /usr/lib/R/lib/libR.so    TRUE
#> 87                               /usr/lib/R/lib/libR.so    TRUE
#> 88                               /usr/lib/R/lib/libR.so    TRUE
#> 89                               /usr/lib/R/lib/libR.so    TRUE
#> 90                               /usr/lib/R/lib/libR.so    TRUE
#> 91                               /usr/lib/R/lib/libR.so    TRUE
#> 92                               /usr/lib/R/lib/libR.so    TRUE
#> 93                               /usr/lib/R/lib/libR.so    TRUE
#> 94                               /usr/lib/R/lib/libR.so    TRUE
#> 95                               /usr/lib/R/lib/libR.so    TRUE
#> 96                               /usr/lib/R/lib/libR.so    TRUE
#> 97                               /usr/lib/R/lib/libR.so    TRUE
#> 98                               /usr/lib/R/lib/libR.so    TRUE
#> 99                               /usr/lib/R/lib/libR.so    TRUE
#> 100                              /usr/lib/R/lib/libR.so    TRUE
#> 101                              /usr/lib/R/lib/libR.so    TRUE
#> 102                              /usr/lib/R/lib/libR.so    TRUE
#> 103                              /usr/lib/R/lib/libR.so    TRUE
#> 104                              /usr/lib/R/lib/libR.so    TRUE
#> 105                              /usr/lib/R/lib/libR.so    TRUE
#> 106                              /usr/lib/R/lib/libR.so    TRUE
#> 107                              /usr/lib/R/lib/libR.so    TRUE
#> 108                              /usr/lib/R/lib/libR.so    TRUE
#> 109                              /usr/lib/R/lib/libR.so    TRUE
#> 110                              /usr/lib/R/lib/libR.so    TRUE
#> 111                              /usr/lib/R/lib/libR.so    TRUE
#> 112                              /usr/lib/R/lib/libR.so    TRUE
#> 113                              /usr/lib/R/lib/libR.so    TRUE
#> 114                              /usr/lib/R/lib/libR.so    TRUE
#> 115                              /usr/lib/R/lib/libR.so    TRUE
#> 116                              /usr/lib/R/lib/libR.so    TRUE
#> 117                              /usr/lib/R/lib/libR.so    TRUE
#> 118                              /usr/lib/R/lib/libR.so    TRUE
#> 119                              /usr/lib/R/lib/libR.so    TRUE
#> 120                              /usr/lib/R/lib/libR.so    TRUE
#> 121                              /usr/lib/R/lib/libR.so    TRUE
#> 122                              /usr/lib/R/lib/libR.so    TRUE
#> 123                              /usr/lib/R/lib/libR.so    TRUE
#> 124                              /usr/lib/R/lib/libR.so    TRUE
#> 125                              /usr/lib/R/lib/libR.so    TRUE
#> 126                              /usr/lib/R/lib/libR.so    TRUE
#> 127                              /usr/lib/R/lib/libR.so    TRUE
#> 128                              /usr/lib/R/lib/libR.so    TRUE
#> 129                              /usr/lib/R/lib/libR.so    TRUE
#> 130                              /usr/lib/R/lib/libR.so    TRUE
#> 131                              /usr/lib/R/lib/libR.so    TRUE
#> 132                              /usr/lib/R/lib/libR.so    TRUE
#> 133                              /usr/lib/R/lib/libR.so    TRUE
#> 134                              /usr/lib/R/lib/libR.so    TRUE
#> 135                              /usr/lib/R/lib/libR.so    TRUE
#> 136                              /usr/lib/R/lib/libR.so    TRUE
#> 137                              /usr/lib/R/lib/libR.so    TRUE
#> 138                              /usr/lib/R/lib/libR.so    TRUE
#> 139                              /usr/lib/R/lib/libR.so    TRUE
#> 140                              /usr/lib/R/lib/libR.so    TRUE
#> 141                              /usr/lib/R/lib/libR.so    TRUE
#> 142                              /usr/lib/R/lib/libR.so    TRUE
#> 143 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 144 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 145                              /usr/lib/R/lib/libR.so    TRUE
#> 146                              /usr/lib/R/lib/libR.so    TRUE
#> 147                              /usr/lib/R/lib/libR.so    TRUE
#> 148 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 149                              /usr/lib/R/lib/libR.so    TRUE
#> 150 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 151 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 152                              /usr/lib/R/lib/libR.so    TRUE
#> 153                              /usr/lib/R/lib/libR.so    TRUE
#> 154                              /usr/lib/R/lib/libR.so    TRUE
#> 155                              /usr/lib/R/lib/libR.so    TRUE
#> 156                              /usr/lib/R/lib/libR.so    TRUE
#> 157                              /usr/lib/R/lib/libR.so    TRUE
#> 158                              /usr/lib/R/lib/libR.so    TRUE
#> 159                              /usr/lib/R/lib/libR.so    TRUE
#> 160                              /usr/lib/R/lib/libR.so    TRUE
#> 161                              /usr/lib/R/lib/libR.so    TRUE
#> 162                              /usr/lib/R/lib/libR.so    TRUE
#> 163                              /usr/lib/R/lib/libR.so    TRUE
#> 164                              /usr/lib/R/lib/libR.so    TRUE
#> 165                              /usr/lib/R/lib/libR.so    TRUE
#> 166                              /usr/lib/R/lib/libR.so    TRUE
#> 167                              /usr/lib/R/lib/libR.so    TRUE
#> 168                              /usr/lib/R/lib/libR.so    TRUE
#> 169                              /usr/lib/R/lib/libR.so    TRUE
#> 170                              /usr/lib/R/lib/libR.so    TRUE
#> 171                              /usr/lib/R/lib/libR.so    TRUE
#> 172                              /usr/lib/R/lib/libR.so    TRUE
#> 173                              /usr/lib/R/lib/libR.so    TRUE
#> 174                              /usr/lib/R/lib/libR.so    TRUE
#> 175                              /usr/lib/R/lib/libR.so    TRUE
#> 176                              /usr/lib/R/lib/libR.so    TRUE
#> 177                              /usr/lib/R/lib/libR.so    TRUE
#> 178                              /usr/lib/R/lib/libR.so    TRUE
#> 179                              /usr/lib/R/lib/libR.so    TRUE
#> 180                              /usr/lib/R/lib/libR.so    TRUE
#> 181                              /usr/lib/R/lib/libR.so    TRUE
#> 182                              /usr/lib/R/lib/libR.so    TRUE
#> 183                              /usr/lib/R/lib/libR.so    TRUE
#> 184                              /usr/lib/R/lib/libR.so    TRUE
#> 185                              /usr/lib/R/lib/libR.so    TRUE
#> 186                              /usr/lib/R/lib/libR.so    TRUE
#> 187                              /usr/lib/R/lib/libR.so    TRUE
#> 188                              /usr/lib/R/lib/libR.so    TRUE
#> 189                              /usr/lib/R/lib/libR.so    TRUE
#> 190                              /usr/lib/R/lib/libR.so    TRUE
#> 191                              /usr/lib/R/lib/libR.so    TRUE
#> 192                              /usr/lib/R/lib/libR.so    TRUE
#> 193                              /usr/lib/R/lib/libR.so    TRUE
#> 194                              /usr/lib/R/lib/libR.so    TRUE
#> 195                              /usr/lib/R/lib/libR.so    TRUE
#> 196                              /usr/lib/R/lib/libR.so    TRUE
#> 197                              /usr/lib/R/lib/libR.so    TRUE
#> 198                              /usr/lib/R/lib/libR.so    TRUE
#> 199                              /usr/lib/R/lib/libR.so    TRUE
#> 200                              /usr/lib/R/lib/libR.so    TRUE
#> 201                              /usr/lib/R/lib/libR.so    TRUE
#> 202                              /usr/lib/R/lib/libR.so    TRUE
#> 203                              /usr/lib/R/lib/libR.so    TRUE
#> 204                              /usr/lib/R/lib/libR.so    TRUE
#> 205                               /usr/lib/R/bin/exec/R    TRUE
#> 206                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 207                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 208                               /usr/lib/R/bin/exec/R    TRUE