Native stack trace

This function returns the native stack trace as a data frame. Each native stack frame corresponds to one row in the returned data frame. Deep function calls come first, the last row corresponds to the running process's entry point.

winch_trace_back()

Value

A data frame with the columns:

func: function name

ip: instruction pointer

pathname: path to shared library

is_libr: a logical, TRUE if this entry is from R's shared library, determined via procmaps::path_is_libr() on the pathname component

Details

On Windows, call winch_init_library() to return function names for a specific package.

See also

Examples

winch_trace_back()
#>                 func               ip
#> 1        R_doDotCall 000072809e90acbd
#> 2        bcEval_loop 000072809e953db3
#> 3             bcEval 000072809e95b27a
#> 4             bcEval 000072809e95b27a
#> 5            Rf_eval 000072809e95b67a
#> 6      R_execClosure 000072809e95d7f6
#> 7  applyClosure_core 000072809e95e5a7
#> 8    Rf_applyClosure 000072809e95b7b1
#> 9            Rf_eval 000072809e95b7b1
#> 10           do_eval 000072809e961fe6
#> 11       bcEval_loop 000072809e94f90c
#> 12            bcEval 000072809e95b27a
#> 13            bcEval 000072809e95b27a
#> 14           Rf_eval 000072809e95b67a
#> 15      forcePromise 000072809e95c034
#> 16      forcePromise 000072809e95b875
#> 17           Rf_eval 000072809e95b875
#> 18    do_withVisible 000072809e96298f
#> 19       do_internal 000072809e9a96a6
#> 20       bcEval_loop 000072809e94f6db
#> 21            bcEval 000072809e95b27a
#> 22            bcEval 000072809e95b27a
#> 23           Rf_eval 000072809e95b67a
#> 24      forcePromise 000072809e95c034
#> 25      forcePromise 000072809e95b875
#> 26           Rf_eval 000072809e95b875
#> 27      forcePromise 000072809e95c034
#> 28      forcePromise 000072809e95c354
#> 29            getvar 000072809e95c354
#> 30       bcEval_loop 000072809e954052
#> 31            bcEval 000072809e95b27a
#> 32            bcEval 000072809e95b27a
#> 33           Rf_eval 000072809e95b67a
#> 34     R_execClosure 000072809e95d7f6
#> 35 applyClosure_core 000072809e95e5a7
#> 36   Rf_applyClosure 000072809e95b7b1
#> 37           Rf_eval 000072809e95b7b1
#> 38           do_eval 000072809e961fe6
#> 39       bcEval_loop 000072809e94f90c
#> 40            bcEval 000072809e95b27a
#> 41            bcEval 000072809e95b27a
#> 42           Rf_eval 000072809e95b67a
#> 43     R_execClosure 000072809e95d7f6
#> 44 applyClosure_core 000072809e95e5a7
#> 45   Rf_applyClosure 000072809e961083
#> 46    R_forceAndCall 000072809e961083
#> 47  0000728078c6eb3e 0000728078c6eb3e
#> 48  0000728078c6ed0d 0000728078c6ed0d
#> 49       R_doDotCall 000072809e90ac29
#> 50        do_dotcall 000072809e90b206
#> 51           Rf_eval 000072809e95bcf5
#> 52  0000728078c6cbcd 0000728078c6cbcd
#> 53 R_ExecWithCleanup 000072809e8f06f3
#> 54  0000728078c6cef2 0000728078c6cef2
#> 55  0000728078c6cf78 0000728078c6cf78
#> 56       R_doDotCall 000072809e90aca9
#> 57       bcEval_loop 000072809e953db3
#> 58            bcEval 000072809e95b27a
#> 59            bcEval 000072809e95b27a
#> 60           Rf_eval 000072809e95b67a
#> 61     R_execClosure 000072809e95d7f6
#> 62 applyClosure_core 000072809e95e5a7
#> 63   Rf_applyClosure 000072809e95b7b1
#> 64           Rf_eval 000072809e95b7b1
#> 65  Rf_ReplIteration 000072809e996934
#> 66     R_ReplConsole 000072809e996d80
#> 67     run_Rmainloop 000072809e996e3f
#> 68  00005c6e8a6e409e 00005c6e8a6e409e
#> 69  000072809e42a8c0 000072809e42a8c0
#> 70 __libc_start_main 000072809e42a9d7
#> 71  00005c6e8a6e40d4 00005c6e8a6e40d4
#> 72  00005c6e8a6e40af 00005c6e8a6e40af
#>                                               pathname is_libr
#> 1                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 2                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 3                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 4                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 5                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 6                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 7                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 8                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 9                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 10                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 11                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 12                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 13                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 14                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 15                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 16                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 17                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 18                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 19                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 20                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 21                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 22                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 23                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 24                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 25                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 26                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 27                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 28                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 29                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 30                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 31                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 32                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 33                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 34                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 35                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 36                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 37                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 38                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 39                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 40                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 41                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 42                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 43                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 44                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 45                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 46                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 47 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 48 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 49                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 50                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 51                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 52 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 53                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 54 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 55 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 56                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 57                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 58                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 59                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 60                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 61                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 62                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 63                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 64                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 65                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 66                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 67                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 68                       /opt/R/4.6.1/lib/R/bin/exec/R    TRUE
#> 69                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 70                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 71                       /opt/R/4.6.1/lib/R/bin/exec/R    TRUE
#> 72                       /opt/R/4.6.1/lib/R/bin/exec/R    TRUE

foo <- function() {
  winch_call(bar)
}

bar <- function() {
  winch_trace_back()
}

foo()
#>                 func               ip
#> 1        R_doDotCall 000072809e90acbd
#> 2        bcEval_loop 000072809e953db3
#> 3             bcEval 000072809e95b27a
#> 4             bcEval 000072809e95b27a
#> 5            Rf_eval 000072809e95b67a
#> 6      R_execClosure 000072809e95d7f6
#> 7  applyClosure_core 000072809e95e5a7
#> 8    Rf_applyClosure 000072809e95b7b1
#> 9            Rf_eval 000072809e95b7b1
#> 10          do_begin 000072809e95f37c
#> 11           Rf_eval 000072809e95ba6f
#> 12     R_execClosure 000072809e95d7f6
#> 13 applyClosure_core 000072809e95e5a7
#> 14   Rf_applyClosure 000072809e95b7b1
#> 15           Rf_eval 000072809e95b7b1
#> 16        winch_call 00007280782508f6
#> 17       R_doDotCall 000072809e90aca9
#> 18       bcEval_loop 000072809e953db3
#> 19            bcEval 000072809e95b27a
#> 20            bcEval 000072809e95b27a
#> 21           Rf_eval 000072809e95b67a
#> 22     R_execClosure 000072809e95d7f6
#> 23 applyClosure_core 000072809e95e5a7
#> 24   Rf_applyClosure 000072809e95b7b1
#> 25           Rf_eval 000072809e95b7b1
#> 26          do_begin 000072809e95f37c
#> 27           Rf_eval 000072809e95ba6f
#> 28     R_execClosure 000072809e95d7f6
#> 29 applyClosure_core 000072809e95e5a7
#> 30   Rf_applyClosure 000072809e95b7b1
#> 31           Rf_eval 000072809e95b7b1
#> 32           do_eval 000072809e961fe6
#> 33       bcEval_loop 000072809e94f90c
#> 34            bcEval 000072809e95b27a
#> 35            bcEval 000072809e95b27a
#> 36           Rf_eval 000072809e95b67a
#> 37      forcePromise 000072809e95c034
#> 38      forcePromise 000072809e95b875
#> 39           Rf_eval 000072809e95b875
#> 40    do_withVisible 000072809e96298f
#> 41       do_internal 000072809e9a96a6
#> 42       bcEval_loop 000072809e94f6db
#> 43            bcEval 000072809e95b27a
#> 44            bcEval 000072809e95b27a
#> 45           Rf_eval 000072809e95b67a
#> 46      forcePromise 000072809e95c034
#> 47      forcePromise 000072809e95b875
#> 48           Rf_eval 000072809e95b875
#> 49      forcePromise 000072809e95c034
#> 50      forcePromise 000072809e95c354
#> 51            getvar 000072809e95c354
#> 52       bcEval_loop 000072809e954052
#> 53            bcEval 000072809e95b27a
#> 54            bcEval 000072809e95b27a
#> 55           Rf_eval 000072809e95b67a
#> 56     R_execClosure 000072809e95d7f6
#> 57 applyClosure_core 000072809e95e5a7
#> 58   Rf_applyClosure 000072809e95b7b1
#> 59           Rf_eval 000072809e95b7b1
#> 60           do_eval 000072809e961fe6
#> 61       bcEval_loop 000072809e94f90c
#> 62            bcEval 000072809e95b27a
#> 63            bcEval 000072809e95b27a
#> 64           Rf_eval 000072809e95b67a
#> 65     R_execClosure 000072809e95d7f6
#> 66 applyClosure_core 000072809e95e5a7
#> 67   Rf_applyClosure 000072809e961083
#> 68    R_forceAndCall 000072809e961083
#> 69  0000728078c6eb3e 0000728078c6eb3e
#> 70  0000728078c6ed0d 0000728078c6ed0d
#> 71       R_doDotCall 000072809e90ac29
#> 72        do_dotcall 000072809e90b206
#> 73           Rf_eval 000072809e95bcf5
#> 74  0000728078c6cbcd 0000728078c6cbcd
#> 75 R_ExecWithCleanup 000072809e8f06f3
#> 76  0000728078c6cef2 0000728078c6cef2
#> 77  0000728078c6cf78 0000728078c6cf78
#> 78       R_doDotCall 000072809e90aca9
#> 79       bcEval_loop 000072809e953db3
#> 80            bcEval 000072809e95b27a
#> 81            bcEval 000072809e95b27a
#> 82           Rf_eval 000072809e95b67a
#> 83     R_execClosure 000072809e95d7f6
#> 84 applyClosure_core 000072809e95e5a7
#> 85   Rf_applyClosure 000072809e95b7b1
#> 86           Rf_eval 000072809e95b7b1
#> 87  Rf_ReplIteration 000072809e996934
#> 88     R_ReplConsole 000072809e996d80
#> 89     run_Rmainloop 000072809e996e3f
#> 90  00005c6e8a6e409e 00005c6e8a6e409e
#> 91  000072809e42a8c0 000072809e42a8c0
#> 92 __libc_start_main 000072809e42a9d7
#> 93  00005c6e8a6e40d4 00005c6e8a6e40d4
#> 94  00005c6e8a6e40af 00005c6e8a6e40af
#>                                               pathname is_libr
#> 1                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 2                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 3                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 4                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 5                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 6                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 7                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 8                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 9                       /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 10                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 11                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 12                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 13                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 14                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 15                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 16 /home/runner/work/_temp/Library/winch/libs/winch.so   FALSE
#> 17                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 18                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 19                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 20                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 21                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 22                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 23                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 24                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 25                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 26                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 27                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 28                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 29                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 30                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 31                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 32                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 33                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 34                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 35                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 36                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 37                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 38                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 39                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 40                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 41                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 42                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 43                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 44                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 45                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 46                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 47                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 48                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 49                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 50                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 51                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 52                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 53                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 54                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 55                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 56                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 57                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 58                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 59                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 60                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 61                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 62                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 63                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 64                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 65                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 66                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 67                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 68                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 69 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 70 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 71                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 72                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 73                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 74 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 75                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 76 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 77 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 78                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 79                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 80                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 81                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 82                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 83                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 84                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 85                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 86                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 87                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 88                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 89                      /opt/R/4.6.1/lib/R/lib/libR.so    TRUE
#> 90                       /opt/R/4.6.1/lib/R/bin/exec/R    TRUE
#> 91                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 92                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 93                       /opt/R/4.6.1/lib/R/bin/exec/R    TRUE
#> 94                       /opt/R/4.6.1/lib/R/bin/exec/R    TRUE