Native stack trace

This function returns the native stack trace as a data frame. Each native stack frame corresponds to one row in the returned data frame. Deep function calls come first, the last row corresponds to the running process's entry point.

winch_trace_back()

Value

A data frame with the columns:

func: function name

ip: instruction pointer

pathname: path to shared library

Details

On Windows, call winch_init_library() to return function names for a specific package.

See also

Examples

winch_trace_back()

#>                   func               ip
#> 1             do_Rprof 00007fd6a9432dc0
#> 2              Rf_eval 00007fd6a9448730
#> 3              Rf_eval 00007fd6a944a3b0
#> 4      Rf_applyClosure 00007fd6a944b2d0
#> 5              Rf_eval 00007fd6a9448730
#> 6       R_forceAndCall 00007fd6a944e5e0
#> 7             do_Rprof 00007fd6a9432dc0
#> 8              Rf_eval 00007fd6a9448730
#> 9              Rf_eval 00007fd6a944a3b0
#> 10     Rf_applyClosure 00007fd6a944b2d0
#> 11            do_Rprof 00007fd6a9432dc0
#> 12             Rf_eval 00007fd6a9448730
#> 13             Rf_eval 00007fd6a94491a0
#> 14             Rf_eval 00007fd6a9448730
#> 15      R_forceAndCall 00007fd6a944f3b0
#> 16         do_Rprofmem 00007fd6a948c320
#> 17            do_Rprof 00007fd6a9432dc0
#> 18             Rf_eval 00007fd6a9448730
#> 19             Rf_eval 00007fd6a944a3b0
#> 20     Rf_applyClosure 00007fd6a944b2d0
#> 21            do_Rprof 00007fd6a9432dc0
#> 22             Rf_eval 00007fd6a9448730
#> 23             Rf_eval 00007fd6a94491a0
#> 24             Rf_eval 00007fd6a94492e0
#> 25            do_Rprof 00007fd6a9432dc0
#> 26             Rf_eval 00007fd6a9448730
#> 27             Rf_eval 00007fd6a944a3b0
#> 28     Rf_applyClosure 00007fd6a944b2d0
#> 29            do_Rprof 00007fd6a9432dc0
#> 30             Rf_eval 00007fd6a9448730
#> 31             Rf_eval 00007fd6a94491a0
#> 32             Rf_eval 00007fd6a94492e0
#> 33            do_Rprof 00007fd6a9432dc0
#> 34             Rf_eval 00007fd6a9448730
#> 35             Rf_eval 00007fd6a94491a0
#> 36             Rf_eval 00007fd6a94492e0
#> 37            do_Rprof 00007fd6a9432dc0
#> 38             Rf_eval 00007fd6a9448730
#> 39             Rf_eval 00007fd6a94491a0
#> 40             Rf_eval 00007fd6a94492e0
#> 41            do_Rprof 00007fd6a9432dc0
#> 42             Rf_eval 00007fd6a9448730
#> 43             Rf_eval 00007fd6a94491a0
#> 44             Rf_eval 00007fd6a94492e0
#> 45            do_Rprof 00007fd6a9432dc0
#> 46             Rf_eval 00007fd6a9448730
#> 47             Rf_eval 00007fd6a94491a0
#> 48             Rf_eval 00007fd6a94492e0
#> 49            do_Rprof 00007fd6a9432dc0
#> 50             Rf_eval 00007fd6a9448730
#> 51             Rf_eval 00007fd6a94491a0
#> 52             Rf_eval 00007fd6a94492e0
#> 53            do_Rprof 00007fd6a9432dc0
#> 54             Rf_eval 00007fd6a9448730
#> 55             Rf_eval 00007fd6a944a3b0
#> 56     Rf_applyClosure 00007fd6a944b2d0
#> 57            do_Rprof 00007fd6a9432dc0
#> 58             Rf_eval 00007fd6a9448730
#> 59             Rf_eval 00007fd6a944a3b0
#> 60     Rf_applyClosure 00007fd6a944b2d0
#> 61            do_Rprof 00007fd6a9432dc0
#> 62             Rf_eval 00007fd6a9448730
#> 63             Rf_eval 00007fd6a944a3b0
#> 64     Rf_applyClosure 00007fd6a944b2d0
#> 65            do_Rprof 00007fd6a9432dc0
#> 66             Rf_eval 00007fd6a9448730
#> 67             Rf_eval 00007fd6a944a3b0
#> 68     Rf_applyClosure 00007fd6a944b2d0
#> 69            do_Rprof 00007fd6a9432dc0
#> 70             Rf_eval 00007fd6a9448730
#> 71             Rf_eval 00007fd6a944a3b0
#> 72     Rf_applyClosure 00007fd6a944b2d0
#> 73            do_Rprof 00007fd6a9432dc0
#> 74             Rf_eval 00007fd6a9448730
#> 75             Rf_eval 00007fd6a944a3b0
#> 76     Rf_applyClosure 00007fd6a944b2d0
#> 77            do_Rprof 00007fd6a9432dc0
#> 78             Rf_eval 00007fd6a9448730
#> 79             Rf_eval 00007fd6a94491a0
#> 80             Rf_eval 00007fd6a94492e0
#> 81            do_Rprof 00007fd6a9432dc0
#> 82             Rf_eval 00007fd6a9448730
#> 83             Rf_eval 00007fd6a944a3b0
#> 84     Rf_applyClosure 00007fd6a944b2d0
#> 85            do_Rprof 00007fd6a9432dc0
#> 86             Rf_eval 00007fd6a9448730
#> 87             Rf_eval 00007fd6a944a3b0
#> 88     Rf_applyClosure 00007fd6a944b2d0
#> 89            do_Rprof 00007fd6a9432dc0
#> 90             Rf_eval 00007fd6a9448730
#> 91             Rf_eval 00007fd6a944a3b0
#> 92     Rf_applyClosure 00007fd6a944b2d0
#> 93            do_Rprof 00007fd6a9432dc0
#> 94             Rf_eval 00007fd6a9448730
#> 95             Rf_eval 00007fd6a944a3b0
#> 96     Rf_applyClosure 00007fd6a944b2d0
#> 97            do_Rprof 00007fd6a9432dc0
#> 98             Rf_eval 00007fd6a9448730
#> 99             Rf_eval 00007fd6a944a3b0
#> 100    Rf_applyClosure 00007fd6a944b2d0
#> 101           do_Rprof 00007fd6a9432dc0
#> 102            Rf_eval 00007fd6a9448730
#> 103            Rf_eval 00007fd6a944a3b0
#> 104    Rf_applyClosure 00007fd6a944b2d0
#> 105           do_Rprof 00007fd6a9432dc0
#> 106            Rf_eval 00007fd6a9448730
#> 107            Rf_eval 00007fd6a944a3b0
#> 108    Rf_applyClosure 00007fd6a944b2d0
#> 109           do_Rprof 00007fd6a9432dc0
#> 110            Rf_eval 00007fd6a9448730
#> 111            Rf_eval 00007fd6a94491a0
#> 112            Rf_eval 00007fd6a94492e0
#> 113           do_Rprof 00007fd6a9432dc0
#> 114            Rf_eval 00007fd6a9448730
#> 115            Rf_eval 00007fd6a944a3b0
#> 116    Rf_applyClosure 00007fd6a944b2d0
#> 117           do_Rprof 00007fd6a9432dc0
#> 118            Rf_eval 00007fd6a9448730
#> 119            Rf_eval 00007fd6a944a3b0
#> 120    Rf_applyClosure 00007fd6a944b2d0
#> 121     R_forceAndCall 00007fd6a944dca0
#> 122          call_loop 00007fd69b4973b0
#> 123           map_impl 00007fd69b497500
#> 124 Rf_NewFrameConfirm 00007fd6a93fca60
#> 125           do_Rprof 00007fd6a9432dc0
#> 126            Rf_eval 00007fd6a9448730
#> 127            Rf_eval 00007fd6a944a3b0
#> 128    Rf_applyClosure 00007fd6a944b2d0
#> 129           do_Rprof 00007fd6a9432dc0
#> 130            Rf_eval 00007fd6a9448730
#> 131            Rf_eval 00007fd6a944a3b0
#> 132    Rf_applyClosure 00007fd6a944b2d0
#> 133           do_Rprof 00007fd6a9432dc0
#> 134            Rf_eval 00007fd6a9448730
#> 135            Rf_eval 00007fd6a944a3b0
#> 136    Rf_applyClosure 00007fd6a944b2d0
#> 137           do_Rprof 00007fd6a9432dc0
#> 138            Rf_eval 00007fd6a9448730
#> 139            Rf_eval 00007fd6a944a3b0
#> 140    Rf_applyClosure 00007fd6a944b2d0
#> 141           do_Rprof 00007fd6a9432dc0
#> 142            Rf_eval 00007fd6a9448730
#> 143            Rf_eval 00007fd6a944a3b0
#> 144    Rf_applyClosure 00007fd6a944b2d0
#> 145            Rf_eval 00007fd6a9448730
#> 146   Rf_ReplIteration 00007fd6a947c3e0
#> 147   Rf_ReplIteration 00007fd6a947c960
#> 148      run_Rmainloop 00007fd6a947ca50
#> 149               main 000055b03f762820
#> 150  __libc_start_main 00007fd6a8f35b10
#> 151             _start 000055b03f762850
#>                                                pathname
#> 1                                /usr/lib/R/lib/libR.so
#> 2                                /usr/lib/R/lib/libR.so
#> 3                                /usr/lib/R/lib/libR.so
#> 4                                /usr/lib/R/lib/libR.so
#> 5                                /usr/lib/R/lib/libR.so
#> 6                                /usr/lib/R/lib/libR.so
#> 7                                /usr/lib/R/lib/libR.so
#> 8                                /usr/lib/R/lib/libR.so
#> 9                                /usr/lib/R/lib/libR.so
#> 10                               /usr/lib/R/lib/libR.so
#> 11                               /usr/lib/R/lib/libR.so
#> 12                               /usr/lib/R/lib/libR.so
#> 13                               /usr/lib/R/lib/libR.so
#> 14                               /usr/lib/R/lib/libR.so
#> 15                               /usr/lib/R/lib/libR.so
#> 16                               /usr/lib/R/lib/libR.so
#> 17                               /usr/lib/R/lib/libR.so
#> 18                               /usr/lib/R/lib/libR.so
#> 19                               /usr/lib/R/lib/libR.so
#> 20                               /usr/lib/R/lib/libR.so
#> 21                               /usr/lib/R/lib/libR.so
#> 22                               /usr/lib/R/lib/libR.so
#> 23                               /usr/lib/R/lib/libR.so
#> 24                               /usr/lib/R/lib/libR.so
#> 25                               /usr/lib/R/lib/libR.so
#> 26                               /usr/lib/R/lib/libR.so
#> 27                               /usr/lib/R/lib/libR.so
#> 28                               /usr/lib/R/lib/libR.so
#> 29                               /usr/lib/R/lib/libR.so
#> 30                               /usr/lib/R/lib/libR.so
#> 31                               /usr/lib/R/lib/libR.so
#> 32                               /usr/lib/R/lib/libR.so
#> 33                               /usr/lib/R/lib/libR.so
#> 34                               /usr/lib/R/lib/libR.so
#> 35                               /usr/lib/R/lib/libR.so
#> 36                               /usr/lib/R/lib/libR.so
#> 37                               /usr/lib/R/lib/libR.so
#> 38                               /usr/lib/R/lib/libR.so
#> 39                               /usr/lib/R/lib/libR.so
#> 40                               /usr/lib/R/lib/libR.so
#> 41                               /usr/lib/R/lib/libR.so
#> 42                               /usr/lib/R/lib/libR.so
#> 43                               /usr/lib/R/lib/libR.so
#> 44                               /usr/lib/R/lib/libR.so
#> 45                               /usr/lib/R/lib/libR.so
#> 46                               /usr/lib/R/lib/libR.so
#> 47                               /usr/lib/R/lib/libR.so
#> 48                               /usr/lib/R/lib/libR.so
#> 49                               /usr/lib/R/lib/libR.so
#> 50                               /usr/lib/R/lib/libR.so
#> 51                               /usr/lib/R/lib/libR.so
#> 52                               /usr/lib/R/lib/libR.so
#> 53                               /usr/lib/R/lib/libR.so
#> 54                               /usr/lib/R/lib/libR.so
#> 55                               /usr/lib/R/lib/libR.so
#> 56                               /usr/lib/R/lib/libR.so
#> 57                               /usr/lib/R/lib/libR.so
#> 58                               /usr/lib/R/lib/libR.so
#> 59                               /usr/lib/R/lib/libR.so
#> 60                               /usr/lib/R/lib/libR.so
#> 61                               /usr/lib/R/lib/libR.so
#> 62                               /usr/lib/R/lib/libR.so
#> 63                               /usr/lib/R/lib/libR.so
#> 64                               /usr/lib/R/lib/libR.so
#> 65                               /usr/lib/R/lib/libR.so
#> 66                               /usr/lib/R/lib/libR.so
#> 67                               /usr/lib/R/lib/libR.so
#> 68                               /usr/lib/R/lib/libR.so
#> 69                               /usr/lib/R/lib/libR.so
#> 70                               /usr/lib/R/lib/libR.so
#> 71                               /usr/lib/R/lib/libR.so
#> 72                               /usr/lib/R/lib/libR.so
#> 73                               /usr/lib/R/lib/libR.so
#> 74                               /usr/lib/R/lib/libR.so
#> 75                               /usr/lib/R/lib/libR.so
#> 76                               /usr/lib/R/lib/libR.so
#> 77                               /usr/lib/R/lib/libR.so
#> 78                               /usr/lib/R/lib/libR.so
#> 79                               /usr/lib/R/lib/libR.so
#> 80                               /usr/lib/R/lib/libR.so
#> 81                               /usr/lib/R/lib/libR.so
#> 82                               /usr/lib/R/lib/libR.so
#> 83                               /usr/lib/R/lib/libR.so
#> 84                               /usr/lib/R/lib/libR.so
#> 85                               /usr/lib/R/lib/libR.so
#> 86                               /usr/lib/R/lib/libR.so
#> 87                               /usr/lib/R/lib/libR.so
#> 88                               /usr/lib/R/lib/libR.so
#> 89                               /usr/lib/R/lib/libR.so
#> 90                               /usr/lib/R/lib/libR.so
#> 91                               /usr/lib/R/lib/libR.so
#> 92                               /usr/lib/R/lib/libR.so
#> 93                               /usr/lib/R/lib/libR.so
#> 94                               /usr/lib/R/lib/libR.so
#> 95                               /usr/lib/R/lib/libR.so
#> 96                               /usr/lib/R/lib/libR.so
#> 97                               /usr/lib/R/lib/libR.so
#> 98                               /usr/lib/R/lib/libR.so
#> 99                               /usr/lib/R/lib/libR.so
#> 100                              /usr/lib/R/lib/libR.so
#> 101                              /usr/lib/R/lib/libR.so
#> 102                              /usr/lib/R/lib/libR.so
#> 103                              /usr/lib/R/lib/libR.so
#> 104                              /usr/lib/R/lib/libR.so
#> 105                              /usr/lib/R/lib/libR.so
#> 106                              /usr/lib/R/lib/libR.so
#> 107                              /usr/lib/R/lib/libR.so
#> 108                              /usr/lib/R/lib/libR.so
#> 109                              /usr/lib/R/lib/libR.so
#> 110                              /usr/lib/R/lib/libR.so
#> 111                              /usr/lib/R/lib/libR.so
#> 112                              /usr/lib/R/lib/libR.so
#> 113                              /usr/lib/R/lib/libR.so
#> 114                              /usr/lib/R/lib/libR.so
#> 115                              /usr/lib/R/lib/libR.so
#> 116                              /usr/lib/R/lib/libR.so
#> 117                              /usr/lib/R/lib/libR.so
#> 118                              /usr/lib/R/lib/libR.so
#> 119                              /usr/lib/R/lib/libR.so
#> 120                              /usr/lib/R/lib/libR.so
#> 121                              /usr/lib/R/lib/libR.so
#> 122 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 123 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 124                              /usr/lib/R/lib/libR.so
#> 125                              /usr/lib/R/lib/libR.so
#> 126                              /usr/lib/R/lib/libR.so
#> 127                              /usr/lib/R/lib/libR.so
#> 128                              /usr/lib/R/lib/libR.so
#> 129                              /usr/lib/R/lib/libR.so
#> 130                              /usr/lib/R/lib/libR.so
#> 131                              /usr/lib/R/lib/libR.so
#> 132                              /usr/lib/R/lib/libR.so
#> 133                              /usr/lib/R/lib/libR.so
#> 134                              /usr/lib/R/lib/libR.so
#> 135                              /usr/lib/R/lib/libR.so
#> 136                              /usr/lib/R/lib/libR.so
#> 137                              /usr/lib/R/lib/libR.so
#> 138                              /usr/lib/R/lib/libR.so
#> 139                              /usr/lib/R/lib/libR.so
#> 140                              /usr/lib/R/lib/libR.so
#> 141                              /usr/lib/R/lib/libR.so
#> 142                              /usr/lib/R/lib/libR.so
#> 143                              /usr/lib/R/lib/libR.so
#> 144                              /usr/lib/R/lib/libR.so
#> 145                              /usr/lib/R/lib/libR.so
#> 146                              /usr/lib/R/lib/libR.so
#> 147                              /usr/lib/R/lib/libR.so
#> 148                              /usr/lib/R/lib/libR.so
#> 149                               /usr/lib/R/bin/exec/R
#> 150                  /lib/x86_64-linux-gnu/libc-2.27.so
#> 151                               /usr/lib/R/bin/exec/R

foo <- function() {
  winch_call(bar)
}

bar <- function() {
  winch_trace_back()
}

foo()

#>                   func               ip
#> 1             do_Rprof 00007fd6a9432dc0
#> 2              Rf_eval 00007fd6a9448730
#> 3              Rf_eval 00007fd6a944a3b0
#> 4      Rf_applyClosure 00007fd6a944b2d0
#> 5              Rf_eval 00007fd6a9448730
#> 6         R_execMethod 00007fd6a944c0d0
#> 7              Rf_eval 00007fd6a9448730
#> 8              Rf_eval 00007fd6a944a3b0
#> 9      Rf_applyClosure 00007fd6a944b2d0
#> 10             Rf_eval 00007fd6a9448730
#> 11          winch_call 00007fd6947a3240
#> 12  Rf_NewFrameConfirm 00007fd6a93fca60
#> 13            do_Rprof 00007fd6a9432dc0
#> 14             Rf_eval 00007fd6a9448730
#> 15             Rf_eval 00007fd6a944a3b0
#> 16     Rf_applyClosure 00007fd6a944b2d0
#> 17             Rf_eval 00007fd6a9448730
#> 18        R_execMethod 00007fd6a944c0d0
#> 19             Rf_eval 00007fd6a9448730
#> 20             Rf_eval 00007fd6a944a3b0
#> 21     Rf_applyClosure 00007fd6a944b2d0
#> 22             Rf_eval 00007fd6a9448730
#> 23      R_forceAndCall 00007fd6a944e5e0
#> 24            do_Rprof 00007fd6a9432dc0
#> 25             Rf_eval 00007fd6a9448730
#> 26             Rf_eval 00007fd6a944a3b0
#> 27     Rf_applyClosure 00007fd6a944b2d0
#> 28            do_Rprof 00007fd6a9432dc0
#> 29             Rf_eval 00007fd6a9448730
#> 30             Rf_eval 00007fd6a94491a0
#> 31             Rf_eval 00007fd6a9448730
#> 32      R_forceAndCall 00007fd6a944f3b0
#> 33         do_Rprofmem 00007fd6a948c320
#> 34            do_Rprof 00007fd6a9432dc0
#> 35             Rf_eval 00007fd6a9448730
#> 36             Rf_eval 00007fd6a944a3b0
#> 37     Rf_applyClosure 00007fd6a944b2d0
#> 38            do_Rprof 00007fd6a9432dc0
#> 39             Rf_eval 00007fd6a9448730
#> 40             Rf_eval 00007fd6a94491a0
#> 41             Rf_eval 00007fd6a94492e0
#> 42            do_Rprof 00007fd6a9432dc0
#> 43             Rf_eval 00007fd6a9448730
#> 44             Rf_eval 00007fd6a944a3b0
#> 45     Rf_applyClosure 00007fd6a944b2d0
#> 46            do_Rprof 00007fd6a9432dc0
#> 47             Rf_eval 00007fd6a9448730
#> 48             Rf_eval 00007fd6a94491a0
#> 49             Rf_eval 00007fd6a94492e0
#> 50            do_Rprof 00007fd6a9432dc0
#> 51             Rf_eval 00007fd6a9448730
#> 52             Rf_eval 00007fd6a94491a0
#> 53             Rf_eval 00007fd6a94492e0
#> 54            do_Rprof 00007fd6a9432dc0
#> 55             Rf_eval 00007fd6a9448730
#> 56             Rf_eval 00007fd6a94491a0
#> 57             Rf_eval 00007fd6a94492e0
#> 58            do_Rprof 00007fd6a9432dc0
#> 59             Rf_eval 00007fd6a9448730
#> 60             Rf_eval 00007fd6a94491a0
#> 61             Rf_eval 00007fd6a94492e0
#> 62            do_Rprof 00007fd6a9432dc0
#> 63             Rf_eval 00007fd6a9448730
#> 64             Rf_eval 00007fd6a94491a0
#> 65             Rf_eval 00007fd6a94492e0
#> 66            do_Rprof 00007fd6a9432dc0
#> 67             Rf_eval 00007fd6a9448730
#> 68             Rf_eval 00007fd6a94491a0
#> 69             Rf_eval 00007fd6a94492e0
#> 70            do_Rprof 00007fd6a9432dc0
#> 71             Rf_eval 00007fd6a9448730
#> 72             Rf_eval 00007fd6a944a3b0
#> 73     Rf_applyClosure 00007fd6a944b2d0
#> 74            do_Rprof 00007fd6a9432dc0
#> 75             Rf_eval 00007fd6a9448730
#> 76             Rf_eval 00007fd6a944a3b0
#> 77     Rf_applyClosure 00007fd6a944b2d0
#> 78            do_Rprof 00007fd6a9432dc0
#> 79             Rf_eval 00007fd6a9448730
#> 80             Rf_eval 00007fd6a944a3b0
#> 81     Rf_applyClosure 00007fd6a944b2d0
#> 82            do_Rprof 00007fd6a9432dc0
#> 83             Rf_eval 00007fd6a9448730
#> 84             Rf_eval 00007fd6a944a3b0
#> 85     Rf_applyClosure 00007fd6a944b2d0
#> 86            do_Rprof 00007fd6a9432dc0
#> 87             Rf_eval 00007fd6a9448730
#> 88             Rf_eval 00007fd6a944a3b0
#> 89     Rf_applyClosure 00007fd6a944b2d0
#> 90            do_Rprof 00007fd6a9432dc0
#> 91             Rf_eval 00007fd6a9448730
#> 92             Rf_eval 00007fd6a944a3b0
#> 93     Rf_applyClosure 00007fd6a944b2d0
#> 94            do_Rprof 00007fd6a9432dc0
#> 95             Rf_eval 00007fd6a9448730
#> 96             Rf_eval 00007fd6a94491a0
#> 97             Rf_eval 00007fd6a94492e0
#> 98            do_Rprof 00007fd6a9432dc0
#> 99             Rf_eval 00007fd6a9448730
#> 100            Rf_eval 00007fd6a944a3b0
#> 101    Rf_applyClosure 00007fd6a944b2d0
#> 102           do_Rprof 00007fd6a9432dc0
#> 103            Rf_eval 00007fd6a9448730
#> 104            Rf_eval 00007fd6a944a3b0
#> 105    Rf_applyClosure 00007fd6a944b2d0
#> 106           do_Rprof 00007fd6a9432dc0
#> 107            Rf_eval 00007fd6a9448730
#> 108            Rf_eval 00007fd6a944a3b0
#> 109    Rf_applyClosure 00007fd6a944b2d0
#> 110           do_Rprof 00007fd6a9432dc0
#> 111            Rf_eval 00007fd6a9448730
#> 112            Rf_eval 00007fd6a944a3b0
#> 113    Rf_applyClosure 00007fd6a944b2d0
#> 114           do_Rprof 00007fd6a9432dc0
#> 115            Rf_eval 00007fd6a9448730
#> 116            Rf_eval 00007fd6a944a3b0
#> 117    Rf_applyClosure 00007fd6a944b2d0
#> 118           do_Rprof 00007fd6a9432dc0
#> 119            Rf_eval 00007fd6a9448730
#> 120            Rf_eval 00007fd6a944a3b0
#> 121    Rf_applyClosure 00007fd6a944b2d0
#> 122           do_Rprof 00007fd6a9432dc0
#> 123            Rf_eval 00007fd6a9448730
#> 124            Rf_eval 00007fd6a944a3b0
#> 125    Rf_applyClosure 00007fd6a944b2d0
#> 126           do_Rprof 00007fd6a9432dc0
#> 127            Rf_eval 00007fd6a9448730
#> 128            Rf_eval 00007fd6a94491a0
#> 129            Rf_eval 00007fd6a94492e0
#> 130           do_Rprof 00007fd6a9432dc0
#> 131            Rf_eval 00007fd6a9448730
#> 132            Rf_eval 00007fd6a944a3b0
#> 133    Rf_applyClosure 00007fd6a944b2d0
#> 134           do_Rprof 00007fd6a9432dc0
#> 135            Rf_eval 00007fd6a9448730
#> 136            Rf_eval 00007fd6a944a3b0
#> 137    Rf_applyClosure 00007fd6a944b2d0
#> 138     R_forceAndCall 00007fd6a944dca0
#> 139          call_loop 00007fd69b4973b0
#> 140           map_impl 00007fd69b497500
#> 141 Rf_NewFrameConfirm 00007fd6a93fca60
#> 142           do_Rprof 00007fd6a9432dc0
#> 143            Rf_eval 00007fd6a9448730
#> 144            Rf_eval 00007fd6a944a3b0
#> 145    Rf_applyClosure 00007fd6a944b2d0
#> 146           do_Rprof 00007fd6a9432dc0
#> 147            Rf_eval 00007fd6a9448730
#> 148            Rf_eval 00007fd6a944a3b0
#> 149    Rf_applyClosure 00007fd6a944b2d0
#> 150           do_Rprof 00007fd6a9432dc0
#> 151            Rf_eval 00007fd6a9448730
#> 152            Rf_eval 00007fd6a944a3b0
#> 153    Rf_applyClosure 00007fd6a944b2d0
#> 154           do_Rprof 00007fd6a9432dc0
#> 155            Rf_eval 00007fd6a9448730
#> 156            Rf_eval 00007fd6a944a3b0
#> 157    Rf_applyClosure 00007fd6a944b2d0
#> 158           do_Rprof 00007fd6a9432dc0
#> 159            Rf_eval 00007fd6a9448730
#> 160            Rf_eval 00007fd6a944a3b0
#> 161    Rf_applyClosure 00007fd6a944b2d0
#> 162            Rf_eval 00007fd6a9448730
#> 163   Rf_ReplIteration 00007fd6a947c3e0
#> 164   Rf_ReplIteration 00007fd6a947c960
#> 165      run_Rmainloop 00007fd6a947ca50
#> 166               main 000055b03f762820
#> 167  __libc_start_main 00007fd6a8f35b10
#> 168             _start 000055b03f762850
#>                                                pathname
#> 1                                /usr/lib/R/lib/libR.so
#> 2                                /usr/lib/R/lib/libR.so
#> 3                                /usr/lib/R/lib/libR.so
#> 4                                /usr/lib/R/lib/libR.so
#> 5                                /usr/lib/R/lib/libR.so
#> 6                                /usr/lib/R/lib/libR.so
#> 7                                /usr/lib/R/lib/libR.so
#> 8                                /usr/lib/R/lib/libR.so
#> 9                                /usr/lib/R/lib/libR.so
#> 10                               /usr/lib/R/lib/libR.so
#> 11  /home/runner/work/_temp/Library/winch/libs/winch.so
#> 12                               /usr/lib/R/lib/libR.so
#> 13                               /usr/lib/R/lib/libR.so
#> 14                               /usr/lib/R/lib/libR.so
#> 15                               /usr/lib/R/lib/libR.so
#> 16                               /usr/lib/R/lib/libR.so
#> 17                               /usr/lib/R/lib/libR.so
#> 18                               /usr/lib/R/lib/libR.so
#> 19                               /usr/lib/R/lib/libR.so
#> 20                               /usr/lib/R/lib/libR.so
#> 21                               /usr/lib/R/lib/libR.so
#> 22                               /usr/lib/R/lib/libR.so
#> 23                               /usr/lib/R/lib/libR.so
#> 24                               /usr/lib/R/lib/libR.so
#> 25                               /usr/lib/R/lib/libR.so
#> 26                               /usr/lib/R/lib/libR.so
#> 27                               /usr/lib/R/lib/libR.so
#> 28                               /usr/lib/R/lib/libR.so
#> 29                               /usr/lib/R/lib/libR.so
#> 30                               /usr/lib/R/lib/libR.so
#> 31                               /usr/lib/R/lib/libR.so
#> 32                               /usr/lib/R/lib/libR.so
#> 33                               /usr/lib/R/lib/libR.so
#> 34                               /usr/lib/R/lib/libR.so
#> 35                               /usr/lib/R/lib/libR.so
#> 36                               /usr/lib/R/lib/libR.so
#> 37                               /usr/lib/R/lib/libR.so
#> 38                               /usr/lib/R/lib/libR.so
#> 39                               /usr/lib/R/lib/libR.so
#> 40                               /usr/lib/R/lib/libR.so
#> 41                               /usr/lib/R/lib/libR.so
#> 42                               /usr/lib/R/lib/libR.so
#> 43                               /usr/lib/R/lib/libR.so
#> 44                               /usr/lib/R/lib/libR.so
#> 45                               /usr/lib/R/lib/libR.so
#> 46                               /usr/lib/R/lib/libR.so
#> 47                               /usr/lib/R/lib/libR.so
#> 48                               /usr/lib/R/lib/libR.so
#> 49                               /usr/lib/R/lib/libR.so
#> 50                               /usr/lib/R/lib/libR.so
#> 51                               /usr/lib/R/lib/libR.so
#> 52                               /usr/lib/R/lib/libR.so
#> 53                               /usr/lib/R/lib/libR.so
#> 54                               /usr/lib/R/lib/libR.so
#> 55                               /usr/lib/R/lib/libR.so
#> 56                               /usr/lib/R/lib/libR.so
#> 57                               /usr/lib/R/lib/libR.so
#> 58                               /usr/lib/R/lib/libR.so
#> 59                               /usr/lib/R/lib/libR.so
#> 60                               /usr/lib/R/lib/libR.so
#> 61                               /usr/lib/R/lib/libR.so
#> 62                               /usr/lib/R/lib/libR.so
#> 63                               /usr/lib/R/lib/libR.so
#> 64                               /usr/lib/R/lib/libR.so
#> 65                               /usr/lib/R/lib/libR.so
#> 66                               /usr/lib/R/lib/libR.so
#> 67                               /usr/lib/R/lib/libR.so
#> 68                               /usr/lib/R/lib/libR.so
#> 69                               /usr/lib/R/lib/libR.so
#> 70                               /usr/lib/R/lib/libR.so
#> 71                               /usr/lib/R/lib/libR.so
#> 72                               /usr/lib/R/lib/libR.so
#> 73                               /usr/lib/R/lib/libR.so
#> 74                               /usr/lib/R/lib/libR.so
#> 75                               /usr/lib/R/lib/libR.so
#> 76                               /usr/lib/R/lib/libR.so
#> 77                               /usr/lib/R/lib/libR.so
#> 78                               /usr/lib/R/lib/libR.so
#> 79                               /usr/lib/R/lib/libR.so
#> 80                               /usr/lib/R/lib/libR.so
#> 81                               /usr/lib/R/lib/libR.so
#> 82                               /usr/lib/R/lib/libR.so
#> 83                               /usr/lib/R/lib/libR.so
#> 84                               /usr/lib/R/lib/libR.so
#> 85                               /usr/lib/R/lib/libR.so
#> 86                               /usr/lib/R/lib/libR.so
#> 87                               /usr/lib/R/lib/libR.so
#> 88                               /usr/lib/R/lib/libR.so
#> 89                               /usr/lib/R/lib/libR.so
#> 90                               /usr/lib/R/lib/libR.so
#> 91                               /usr/lib/R/lib/libR.so
#> 92                               /usr/lib/R/lib/libR.so
#> 93                               /usr/lib/R/lib/libR.so
#> 94                               /usr/lib/R/lib/libR.so
#> 95                               /usr/lib/R/lib/libR.so
#> 96                               /usr/lib/R/lib/libR.so
#> 97                               /usr/lib/R/lib/libR.so
#> 98                               /usr/lib/R/lib/libR.so
#> 99                               /usr/lib/R/lib/libR.so
#> 100                              /usr/lib/R/lib/libR.so
#> 101                              /usr/lib/R/lib/libR.so
#> 102                              /usr/lib/R/lib/libR.so
#> 103                              /usr/lib/R/lib/libR.so
#> 104                              /usr/lib/R/lib/libR.so
#> 105                              /usr/lib/R/lib/libR.so
#> 106                              /usr/lib/R/lib/libR.so
#> 107                              /usr/lib/R/lib/libR.so
#> 108                              /usr/lib/R/lib/libR.so
#> 109                              /usr/lib/R/lib/libR.so
#> 110                              /usr/lib/R/lib/libR.so
#> 111                              /usr/lib/R/lib/libR.so
#> 112                              /usr/lib/R/lib/libR.so
#> 113                              /usr/lib/R/lib/libR.so
#> 114                              /usr/lib/R/lib/libR.so
#> 115                              /usr/lib/R/lib/libR.so
#> 116                              /usr/lib/R/lib/libR.so
#> 117                              /usr/lib/R/lib/libR.so
#> 118                              /usr/lib/R/lib/libR.so
#> 119                              /usr/lib/R/lib/libR.so
#> 120                              /usr/lib/R/lib/libR.so
#> 121                              /usr/lib/R/lib/libR.so
#> 122                              /usr/lib/R/lib/libR.so
#> 123                              /usr/lib/R/lib/libR.so
#> 124                              /usr/lib/R/lib/libR.so
#> 125                              /usr/lib/R/lib/libR.so
#> 126                              /usr/lib/R/lib/libR.so
#> 127                              /usr/lib/R/lib/libR.so
#> 128                              /usr/lib/R/lib/libR.so
#> 129                              /usr/lib/R/lib/libR.so
#> 130                              /usr/lib/R/lib/libR.so
#> 131                              /usr/lib/R/lib/libR.so
#> 132                              /usr/lib/R/lib/libR.so
#> 133                              /usr/lib/R/lib/libR.so
#> 134                              /usr/lib/R/lib/libR.so
#> 135                              /usr/lib/R/lib/libR.so
#> 136                              /usr/lib/R/lib/libR.so
#> 137                              /usr/lib/R/lib/libR.so
#> 138                              /usr/lib/R/lib/libR.so
#> 139 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 140 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 141                              /usr/lib/R/lib/libR.so
#> 142                              /usr/lib/R/lib/libR.so
#> 143                              /usr/lib/R/lib/libR.so
#> 144                              /usr/lib/R/lib/libR.so
#> 145                              /usr/lib/R/lib/libR.so
#> 146                              /usr/lib/R/lib/libR.so
#> 147                              /usr/lib/R/lib/libR.so
#> 148                              /usr/lib/R/lib/libR.so
#> 149                              /usr/lib/R/lib/libR.so
#> 150                              /usr/lib/R/lib/libR.so
#> 151                              /usr/lib/R/lib/libR.so
#> 152                              /usr/lib/R/lib/libR.so
#> 153                              /usr/lib/R/lib/libR.so
#> 154                              /usr/lib/R/lib/libR.so
#> 155                              /usr/lib/R/lib/libR.so
#> 156                              /usr/lib/R/lib/libR.so
#> 157                              /usr/lib/R/lib/libR.so
#> 158                              /usr/lib/R/lib/libR.so
#> 159                              /usr/lib/R/lib/libR.so
#> 160                              /usr/lib/R/lib/libR.so
#> 161                              /usr/lib/R/lib/libR.so
#> 162                              /usr/lib/R/lib/libR.so
#> 163                              /usr/lib/R/lib/libR.so
#> 164                              /usr/lib/R/lib/libR.so
#> 165                              /usr/lib/R/lib/libR.so
#> 166                               /usr/lib/R/bin/exec/R
#> 167                  /lib/x86_64-linux-gnu/libc-2.27.so
#> 168                               /usr/lib/R/bin/exec/R