Native stack trace

This function returns the native stack trace as a data frame. Each native stack frame corresponds to one row in the returned data frame. Deep function calls come first, the last row corresponds to the running process's entry point.

winch_trace_back()

Value

A data frame with the columns:

func: function name

ip: instruction pointer

pathname: path to shared library

is_libr: a logical, TRUE if this entry is from R's shared library, determined via procmaps::path_is_libr() on the pathname component

Details

On Windows, call winch_init_library() to return function names for a specific package.

See also

Examples

winch_trace_back()
#>                                func               ip
#> 1                     R_ParseString 00007f0fa7d35f60
#> 2                     R_ParseString 00007f0fa7d53f50
#> 3                           Rf_eval 00007f0fa7d54390
#> 4                           Rf_eval 00007f0fa7d56410
#> 5                           Rf_eval 00007f0fa7d57440
#> 6                           Rf_eval 00007f0fa7d54390
#> 7                    R_forceAndCall 00007f0fa7d5a880
#> 8                     R_ParseString 00007f0fa7d35f60
#> 9                     R_ParseString 00007f0fa7d53f50
#> 10                          Rf_eval 00007f0fa7d54390
#> 11                          Rf_eval 00007f0fa7d54df0
#> 12                          Rf_eval 00007f0fa7d54390
#> 13                   R_forceAndCall 00007f0fa7d5b340
#> 14                      do_Rprofmem 00007f0fa7d9ae80
#> 15                    R_ParseString 00007f0fa7d35f60
#> 16                    R_ParseString 00007f0fa7d53f50
#> 17                          Rf_eval 00007f0fa7d54390
#> 18                          Rf_eval 00007f0fa7d56410
#> 19                          Rf_eval 00007f0fa7d57440
#> 20                   R_forceAndCall 00007f0fa7d59b80
#> 21                     R_init_purrr 00007f0fa43d9bf0
#> 22                     R_init_purrr 00007f0fa43d9e00
#> 23               Rf_NewFrameConfirm 00007f0fa7cfb950
#> 24               Rf_NewFrameConfirm 00007f0fa7cfd4d0
#> 25                          Rf_eval 00007f0fa7d54390
#> 26 <unw_get_proc_name() error: -10> 00007f0fa43d7c20
#> 27                R_ExecWithCleanup 00007f0fa7ce42b0
#> 28 <unw_get_proc_name() error: -10> 00007f0fa43d7eb0
#> 29 <unw_get_proc_name() error: -10> 00007f0fa43d7f60
#> 30               Rf_NewFrameConfirm 00007f0fa7cfb950
#> 31                    R_ParseString 00007f0fa7d35f60
#> 32                    R_ParseString 00007f0fa7d53f50
#> 33                          Rf_eval 00007f0fa7d54390
#> 34                          Rf_eval 00007f0fa7d56410
#> 35                          Rf_eval 00007f0fa7d57440
#> 36                          Rf_eval 00007f0fa7d54390
#> 37                 Rf_ReplIteration 00007f0fa7d8aae0
#> 38                 Rf_ReplIteration 00007f0fa7d8aff0
#> 39                    run_Rmainloop 00007f0fa7d8b0f0
#> 40                             main 0000555797fa0080
#> 41                __libc_init_first 00007f0fa7829d10
#> 42                __libc_start_main 00007f0fa7829dc0
#> 43                           _start 0000555797fa00b0
#>                                               pathname is_libr
#> 1                               /usr/lib/R/lib/libR.so    TRUE
#> 2                               /usr/lib/R/lib/libR.so    TRUE
#> 3                               /usr/lib/R/lib/libR.so    TRUE
#> 4                               /usr/lib/R/lib/libR.so    TRUE
#> 5                               /usr/lib/R/lib/libR.so    TRUE
#> 6                               /usr/lib/R/lib/libR.so    TRUE
#> 7                               /usr/lib/R/lib/libR.so    TRUE
#> 8                               /usr/lib/R/lib/libR.so    TRUE
#> 9                               /usr/lib/R/lib/libR.so    TRUE
#> 10                              /usr/lib/R/lib/libR.so    TRUE
#> 11                              /usr/lib/R/lib/libR.so    TRUE
#> 12                              /usr/lib/R/lib/libR.so    TRUE
#> 13                              /usr/lib/R/lib/libR.so    TRUE
#> 14                              /usr/lib/R/lib/libR.so    TRUE
#> 15                              /usr/lib/R/lib/libR.so    TRUE
#> 16                              /usr/lib/R/lib/libR.so    TRUE
#> 17                              /usr/lib/R/lib/libR.so    TRUE
#> 18                              /usr/lib/R/lib/libR.so    TRUE
#> 19                              /usr/lib/R/lib/libR.so    TRUE
#> 20                              /usr/lib/R/lib/libR.so    TRUE
#> 21 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 22 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 23                              /usr/lib/R/lib/libR.so    TRUE
#> 24                              /usr/lib/R/lib/libR.so    TRUE
#> 25                              /usr/lib/R/lib/libR.so    TRUE
#> 26 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 27                              /usr/lib/R/lib/libR.so    TRUE
#> 28 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 29 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 30                              /usr/lib/R/lib/libR.so    TRUE
#> 31                              /usr/lib/R/lib/libR.so    TRUE
#> 32                              /usr/lib/R/lib/libR.so    TRUE
#> 33                              /usr/lib/R/lib/libR.so    TRUE
#> 34                              /usr/lib/R/lib/libR.so    TRUE
#> 35                              /usr/lib/R/lib/libR.so    TRUE
#> 36                              /usr/lib/R/lib/libR.so    TRUE
#> 37                              /usr/lib/R/lib/libR.so    TRUE
#> 38                              /usr/lib/R/lib/libR.so    TRUE
#> 39                              /usr/lib/R/lib/libR.so    TRUE
#> 40                               /usr/lib/R/bin/exec/R    TRUE
#> 41                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 42                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 43                               /usr/lib/R/bin/exec/R    TRUE

foo <- function() {
  winch_call(bar)
}

bar <- function() {
  winch_trace_back()
}

foo()
#>                                func               ip
#> 1                     R_ParseString 00007f0fa7d35f60
#> 2                     R_ParseString 00007f0fa7d53f50
#> 3                           Rf_eval 00007f0fa7d54390
#> 4                           Rf_eval 00007f0fa7d56410
#> 5                           Rf_eval 00007f0fa7d57440
#> 6                           Rf_eval 00007f0fa7d54390
#> 7                      R_execMethod 00007f0fa7d58160
#> 8                           Rf_eval 00007f0fa7d54390
#> 9                           Rf_eval 00007f0fa7d56410
#> 10                          Rf_eval 00007f0fa7d57440
#> 11                          Rf_eval 00007f0fa7d54390
#> 12                       winch_call 00007f0f9c557710
#> 13               Rf_NewFrameConfirm 00007f0fa7cfb950
#> 14                    R_ParseString 00007f0fa7d35f60
#> 15                    R_ParseString 00007f0fa7d53f50
#> 16                          Rf_eval 00007f0fa7d54390
#> 17                          Rf_eval 00007f0fa7d56410
#> 18                          Rf_eval 00007f0fa7d57440
#> 19                          Rf_eval 00007f0fa7d54390
#> 20                     R_execMethod 00007f0fa7d58160
#> 21                          Rf_eval 00007f0fa7d54390
#> 22                          Rf_eval 00007f0fa7d56410
#> 23                          Rf_eval 00007f0fa7d57440
#> 24                          Rf_eval 00007f0fa7d54390
#> 25                   R_forceAndCall 00007f0fa7d5a880
#> 26                    R_ParseString 00007f0fa7d35f60
#> 27                    R_ParseString 00007f0fa7d53f50
#> 28                          Rf_eval 00007f0fa7d54390
#> 29                          Rf_eval 00007f0fa7d54df0
#> 30                          Rf_eval 00007f0fa7d54390
#> 31                   R_forceAndCall 00007f0fa7d5b340
#> 32                      do_Rprofmem 00007f0fa7d9ae80
#> 33                    R_ParseString 00007f0fa7d35f60
#> 34                    R_ParseString 00007f0fa7d53f50
#> 35                          Rf_eval 00007f0fa7d54390
#> 36                          Rf_eval 00007f0fa7d56410
#> 37                          Rf_eval 00007f0fa7d57440
#> 38                   R_forceAndCall 00007f0fa7d59b80
#> 39                     R_init_purrr 00007f0fa43d9bf0
#> 40                     R_init_purrr 00007f0fa43d9e00
#> 41               Rf_NewFrameConfirm 00007f0fa7cfb950
#> 42               Rf_NewFrameConfirm 00007f0fa7cfd4d0
#> 43                          Rf_eval 00007f0fa7d54390
#> 44 <unw_get_proc_name() error: -10> 00007f0fa43d7c20
#> 45                R_ExecWithCleanup 00007f0fa7ce42b0
#> 46 <unw_get_proc_name() error: -10> 00007f0fa43d7eb0
#> 47 <unw_get_proc_name() error: -10> 00007f0fa43d7f60
#> 48               Rf_NewFrameConfirm 00007f0fa7cfb950
#> 49                    R_ParseString 00007f0fa7d35f60
#> 50                    R_ParseString 00007f0fa7d53f50
#> 51                          Rf_eval 00007f0fa7d54390
#> 52                          Rf_eval 00007f0fa7d56410
#> 53                          Rf_eval 00007f0fa7d57440
#> 54                          Rf_eval 00007f0fa7d54390
#> 55                 Rf_ReplIteration 00007f0fa7d8aae0
#> 56                 Rf_ReplIteration 00007f0fa7d8aff0
#> 57                    run_Rmainloop 00007f0fa7d8b0f0
#> 58                             main 0000555797fa0080
#> 59                __libc_init_first 00007f0fa7829d10
#> 60                __libc_start_main 00007f0fa7829dc0
#> 61                           _start 0000555797fa00b0
#>                                               pathname is_libr
#> 1                               /usr/lib/R/lib/libR.so    TRUE
#> 2                               /usr/lib/R/lib/libR.so    TRUE
#> 3                               /usr/lib/R/lib/libR.so    TRUE
#> 4                               /usr/lib/R/lib/libR.so    TRUE
#> 5                               /usr/lib/R/lib/libR.so    TRUE
#> 6                               /usr/lib/R/lib/libR.so    TRUE
#> 7                               /usr/lib/R/lib/libR.so    TRUE
#> 8                               /usr/lib/R/lib/libR.so    TRUE
#> 9                               /usr/lib/R/lib/libR.so    TRUE
#> 10                              /usr/lib/R/lib/libR.so    TRUE
#> 11                              /usr/lib/R/lib/libR.so    TRUE
#> 12 /home/runner/work/_temp/Library/winch/libs/winch.so   FALSE
#> 13                              /usr/lib/R/lib/libR.so    TRUE
#> 14                              /usr/lib/R/lib/libR.so    TRUE
#> 15                              /usr/lib/R/lib/libR.so    TRUE
#> 16                              /usr/lib/R/lib/libR.so    TRUE
#> 17                              /usr/lib/R/lib/libR.so    TRUE
#> 18                              /usr/lib/R/lib/libR.so    TRUE
#> 19                              /usr/lib/R/lib/libR.so    TRUE
#> 20                              /usr/lib/R/lib/libR.so    TRUE
#> 21                              /usr/lib/R/lib/libR.so    TRUE
#> 22                              /usr/lib/R/lib/libR.so    TRUE
#> 23                              /usr/lib/R/lib/libR.so    TRUE
#> 24                              /usr/lib/R/lib/libR.so    TRUE
#> 25                              /usr/lib/R/lib/libR.so    TRUE
#> 26                              /usr/lib/R/lib/libR.so    TRUE
#> 27                              /usr/lib/R/lib/libR.so    TRUE
#> 28                              /usr/lib/R/lib/libR.so    TRUE
#> 29                              /usr/lib/R/lib/libR.so    TRUE
#> 30                              /usr/lib/R/lib/libR.so    TRUE
#> 31                              /usr/lib/R/lib/libR.so    TRUE
#> 32                              /usr/lib/R/lib/libR.so    TRUE
#> 33                              /usr/lib/R/lib/libR.so    TRUE
#> 34                              /usr/lib/R/lib/libR.so    TRUE
#> 35                              /usr/lib/R/lib/libR.so    TRUE
#> 36                              /usr/lib/R/lib/libR.so    TRUE
#> 37                              /usr/lib/R/lib/libR.so    TRUE
#> 38                              /usr/lib/R/lib/libR.so    TRUE
#> 39 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 40 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 41                              /usr/lib/R/lib/libR.so    TRUE
#> 42                              /usr/lib/R/lib/libR.so    TRUE
#> 43                              /usr/lib/R/lib/libR.so    TRUE
#> 44 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 45                              /usr/lib/R/lib/libR.so    TRUE
#> 46 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 47 /home/runner/work/_temp/Library/purrr/libs/purrr.so   FALSE
#> 48                              /usr/lib/R/lib/libR.so    TRUE
#> 49                              /usr/lib/R/lib/libR.so    TRUE
#> 50                              /usr/lib/R/lib/libR.so    TRUE
#> 51                              /usr/lib/R/lib/libR.so    TRUE
#> 52                              /usr/lib/R/lib/libR.so    TRUE
#> 53                              /usr/lib/R/lib/libR.so    TRUE
#> 54                              /usr/lib/R/lib/libR.so    TRUE
#> 55                              /usr/lib/R/lib/libR.so    TRUE
#> 56                              /usr/lib/R/lib/libR.so    TRUE
#> 57                              /usr/lib/R/lib/libR.so    TRUE
#> 58                               /usr/lib/R/bin/exec/R    TRUE
#> 59                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 60                 /usr/lib/x86_64-linux-gnu/libc.so.6   FALSE
#> 61                               /usr/lib/R/bin/exec/R    TRUE