Native stack trace

This function returns the native stack trace as a data frame. Each native stack frame corresponds to one row in the returned data frame. Deep function calls come first, the last row corresponds to the running process's entry point.

winch_trace_back()

Value

A data frame with the columns:

func: function name

ip: instruction pointer

pathname: path to shared library

Details

On Windows, call winch_init_library() to return function names for a specific package.

See also

Examples

winch_trace_back()

#>                  func               ip
#> 1              bcEval 00007f57815ab2f0
#> 2             Rf_eval 00007f57815c0c60
#> 3       R_execClosure 00007f57815c28e0
#> 4     Rf_applyClosure 00007f57815c3800
#> 5             Rf_eval 00007f57815c0c60
#> 6             do_eval 00007f57815c6b10
#> 7              bcEval 00007f57815ab2f0
#> 8             Rf_eval 00007f57815c0c60
#> 9       R_execClosure 00007f57815c28e0
#> 10    Rf_applyClosure 00007f57815c3800
#> 11             bcEval 00007f57815ab2f0
#> 12            Rf_eval 00007f57815c0c60
#> 13       forcePromise 00007f57815c16d0
#> 14            Rf_eval 00007f57815c0c60
#> 15     do_withVisible 00007f57815c78e0
#> 16        do_internal 00007f57816047a0
#> 17             bcEval 00007f57815ab2f0
#> 18            Rf_eval 00007f57815c0c60
#> 19      R_execClosure 00007f57815c28e0
#> 20    Rf_applyClosure 00007f57815c3800
#> 21             bcEval 00007f57815ab2f0
#> 22            Rf_eval 00007f57815c0c60
#> 23       forcePromise 00007f57815c16d0
#> 24             getvar 00007f57815c1810
#> 25             bcEval 00007f57815ab2f0
#> 26            Rf_eval 00007f57815c0c60
#> 27      R_execClosure 00007f57815c28e0
#> 28    Rf_applyClosure 00007f57815c3800
#> 29             bcEval 00007f57815ab2f0
#> 30            Rf_eval 00007f57815c0c60
#> 31       forcePromise 00007f57815c16d0
#> 32             getvar 00007f57815c1810
#> 33             bcEval 00007f57815ab2f0
#> 34            Rf_eval 00007f57815c0c60
#> 35       forcePromise 00007f57815c16d0
#> 36             getvar 00007f57815c1810
#> 37             bcEval 00007f57815ab2f0
#> 38            Rf_eval 00007f57815c0c60
#> 39       forcePromise 00007f57815c16d0
#> 40             getvar 00007f57815c1810
#> 41             bcEval 00007f57815ab2f0
#> 42            Rf_eval 00007f57815c0c60
#> 43       forcePromise 00007f57815c16d0
#> 44             getvar 00007f57815c1810
#> 45             bcEval 00007f57815ab2f0
#> 46            Rf_eval 00007f57815c0c60
#> 47       forcePromise 00007f57815c16d0
#> 48             getvar 00007f57815c1810
#> 49             bcEval 00007f57815ab2f0
#> 50            Rf_eval 00007f57815c0c60
#> 51       forcePromise 00007f57815c16d0
#> 52             getvar 00007f57815c1810
#> 53             bcEval 00007f57815ab2f0
#> 54            Rf_eval 00007f57815c0c60
#> 55      R_execClosure 00007f57815c28e0
#> 56    Rf_applyClosure 00007f57815c3800
#> 57             bcEval 00007f57815ab2f0
#> 58            Rf_eval 00007f57815c0c60
#> 59      R_execClosure 00007f57815c28e0
#> 60    Rf_applyClosure 00007f57815c3800
#> 61             bcEval 00007f57815ab2f0
#> 62            Rf_eval 00007f57815c0c60
#> 63      R_execClosure 00007f57815c28e0
#> 64    Rf_applyClosure 00007f57815c3800
#> 65             bcEval 00007f57815ab2f0
#> 66            Rf_eval 00007f57815c0c60
#> 67      R_execClosure 00007f57815c28e0
#> 68    Rf_applyClosure 00007f57815c3800
#> 69             bcEval 00007f57815ab2f0
#> 70            Rf_eval 00007f57815c0c60
#> 71      R_execClosure 00007f57815c28e0
#> 72    Rf_applyClosure 00007f57815c3800
#> 73             bcEval 00007f57815ab2f0
#> 74            Rf_eval 00007f57815c0c60
#> 75      R_execClosure 00007f57815c28e0
#> 76    Rf_applyClosure 00007f57815c3800
#> 77             bcEval 00007f57815ab2f0
#> 78            Rf_eval 00007f57815c0c60
#> 79       forcePromise 00007f57815c16d0
#> 80             getvar 00007f57815c1810
#> 81             bcEval 00007f57815ab2f0
#> 82            Rf_eval 00007f57815c0c60
#> 83      R_execClosure 00007f57815c28e0
#> 84    Rf_applyClosure 00007f57815c3800
#> 85             bcEval 00007f57815ab2f0
#> 86            Rf_eval 00007f57815c0c60
#> 87      R_execClosure 00007f57815c28e0
#> 88    Rf_applyClosure 00007f57815c3800
#> 89             bcEval 00007f57815ab2f0
#> 90            Rf_eval 00007f57815c0c60
#> 91      R_execClosure 00007f57815c28e0
#> 92    Rf_applyClosure 00007f57815c3800
#> 93             bcEval 00007f57815ab2f0
#> 94            Rf_eval 00007f57815c0c60
#> 95      R_execClosure 00007f57815c28e0
#> 96    Rf_applyClosure 00007f57815c3800
#> 97             bcEval 00007f57815ab2f0
#> 98            Rf_eval 00007f57815c0c60
#> 99      R_execClosure 00007f57815c28e0
#> 100   Rf_applyClosure 00007f57815c3800
#> 101            bcEval 00007f57815ab2f0
#> 102           Rf_eval 00007f57815c0c60
#> 103     R_execClosure 00007f57815c28e0
#> 104   Rf_applyClosure 00007f57815c3800
#> 105            bcEval 00007f57815ab2f0
#> 106           Rf_eval 00007f57815c0c60
#> 107     R_execClosure 00007f57815c28e0
#> 108   Rf_applyClosure 00007f57815c3800
#> 109            bcEval 00007f57815ab2f0
#> 110           Rf_eval 00007f57815c0c60
#> 111      forcePromise 00007f57815c16d0
#> 112            getvar 00007f57815c1810
#> 113            bcEval 00007f57815ab2f0
#> 114           Rf_eval 00007f57815c0c60
#> 115     R_execClosure 00007f57815c28e0
#> 116   Rf_applyClosure 00007f57815c3800
#> 117            bcEval 00007f57815ab2f0
#> 118           Rf_eval 00007f57815c0c60
#> 119     R_execClosure 00007f57815c28e0
#> 120   Rf_applyClosure 00007f57815c3800
#> 121    R_forceAndCall 00007f57815c61d0
#> 122         call_loop 00007f576fcf43b0
#> 123          map_impl 00007f576fcf4500
#> 124       R_doDotCall 00007f5781575010
#> 125            bcEval 00007f57815ab2f0
#> 126           Rf_eval 00007f57815c0c60
#> 127     R_execClosure 00007f57815c28e0
#> 128   Rf_applyClosure 00007f57815c3800
#> 129            bcEval 00007f57815ab2f0
#> 130           Rf_eval 00007f57815c0c60
#> 131     R_execClosure 00007f57815c28e0
#> 132   Rf_applyClosure 00007f57815c3800
#> 133            bcEval 00007f57815ab2f0
#> 134           Rf_eval 00007f57815c0c60
#> 135     R_execClosure 00007f57815c28e0
#> 136   Rf_applyClosure 00007f57815c3800
#> 137            bcEval 00007f57815ab2f0
#> 138           Rf_eval 00007f57815c0c60
#> 139     R_execClosure 00007f57815c28e0
#> 140   Rf_applyClosure 00007f57815c3800
#> 141            bcEval 00007f57815ab2f0
#> 142           Rf_eval 00007f57815c0c60
#> 143     R_execClosure 00007f57815c28e0
#> 144   Rf_applyClosure 00007f57815c3800
#> 145           Rf_eval 00007f57815c0c60
#> 146  Rf_ReplIteration 00007f57815f4900
#> 147     R_ReplConsole 00007f57815f4e80
#> 148     run_Rmainloop 00007f57815f4f70
#> 149              main 0000562713ebd820
#> 150 __libc_start_main 00007f57810a7b10
#> 151            _start 0000562713ebd850
#>                                                pathname
#> 1                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 2                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 3                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 4                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 5                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 6                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 7                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 8                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 9                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 10                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 11                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 12                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 13                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 14                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 15                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 16                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 17                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 18                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 19                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 20                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 21                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 22                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 23                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 24                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 25                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 26                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 27                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 28                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 29                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 30                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 31                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 32                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 33                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 34                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 35                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 36                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 37                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 38                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 39                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 40                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 41                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 42                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 43                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 44                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 45                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 46                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 47                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 48                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 49                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 50                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 51                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 52                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 53                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 54                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 55                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 56                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 57                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 58                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 59                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 60                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 61                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 62                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 63                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 64                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 65                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 66                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 67                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 68                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 69                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 70                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 71                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 72                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 73                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 74                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 75                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 76                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 77                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 78                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 79                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 80                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 81                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 82                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 83                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 84                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 85                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 86                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 87                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 88                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 89                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 90                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 91                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 92                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 93                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 94                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 95                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 96                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 97                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 98                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 99                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 100                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 101                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 102                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 103                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 104                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 105                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 106                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 107                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 108                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 109                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 110                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 111                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 112                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 113                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 114                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 115                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 116                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 117                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 118                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 119                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 120                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 121                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 122 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 123 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 124                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 125                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 126                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 127                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 128                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 129                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 130                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 131                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 132                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 133                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 134                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 135                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 136                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 137                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 138                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 139                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 140                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 141                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 142                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 143                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 144                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 145                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 146                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 147                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 148                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 149                       /opt/R/4.0.3/lib/R/bin/exec/R
#> 150                  /lib/x86_64-linux-gnu/libc-2.27.so
#> 151                       /opt/R/4.0.3/lib/R/bin/exec/R

foo <- function() {
  winch_call(bar)
}

bar <- function() {
  winch_trace_back()
}

foo()

#>                  func               ip
#> 1              bcEval 00007f57815ab2f0
#> 2             Rf_eval 00007f57815c0c60
#> 3       R_execClosure 00007f57815c28e0
#> 4     Rf_applyClosure 00007f57815c3800
#> 5             Rf_eval 00007f57815c0c60
#> 6            do_begin 00007f57815c4600
#> 7             Rf_eval 00007f57815c0c60
#> 8       R_execClosure 00007f57815c28e0
#> 9     Rf_applyClosure 00007f57815c3800
#> 10            Rf_eval 00007f57815c0c60
#> 11         winch_call 00007f57691e05b0
#> 12        R_doDotCall 00007f5781575010
#> 13             bcEval 00007f57815ab2f0
#> 14            Rf_eval 00007f57815c0c60
#> 15      R_execClosure 00007f57815c28e0
#> 16    Rf_applyClosure 00007f57815c3800
#> 17            Rf_eval 00007f57815c0c60
#> 18           do_begin 00007f57815c4600
#> 19            Rf_eval 00007f57815c0c60
#> 20      R_execClosure 00007f57815c28e0
#> 21    Rf_applyClosure 00007f57815c3800
#> 22            Rf_eval 00007f57815c0c60
#> 23            do_eval 00007f57815c6b10
#> 24             bcEval 00007f57815ab2f0
#> 25            Rf_eval 00007f57815c0c60
#> 26      R_execClosure 00007f57815c28e0
#> 27    Rf_applyClosure 00007f57815c3800
#> 28             bcEval 00007f57815ab2f0
#> 29            Rf_eval 00007f57815c0c60
#> 30       forcePromise 00007f57815c16d0
#> 31            Rf_eval 00007f57815c0c60
#> 32     do_withVisible 00007f57815c78e0
#> 33        do_internal 00007f57816047a0
#> 34             bcEval 00007f57815ab2f0
#> 35            Rf_eval 00007f57815c0c60
#> 36      R_execClosure 00007f57815c28e0
#> 37    Rf_applyClosure 00007f57815c3800
#> 38             bcEval 00007f57815ab2f0
#> 39            Rf_eval 00007f57815c0c60
#> 40       forcePromise 00007f57815c16d0
#> 41             getvar 00007f57815c1810
#> 42             bcEval 00007f57815ab2f0
#> 43            Rf_eval 00007f57815c0c60
#> 44      R_execClosure 00007f57815c28e0
#> 45    Rf_applyClosure 00007f57815c3800
#> 46             bcEval 00007f57815ab2f0
#> 47            Rf_eval 00007f57815c0c60
#> 48       forcePromise 00007f57815c16d0
#> 49             getvar 00007f57815c1810
#> 50             bcEval 00007f57815ab2f0
#> 51            Rf_eval 00007f57815c0c60
#> 52       forcePromise 00007f57815c16d0
#> 53             getvar 00007f57815c1810
#> 54             bcEval 00007f57815ab2f0
#> 55            Rf_eval 00007f57815c0c60
#> 56       forcePromise 00007f57815c16d0
#> 57             getvar 00007f57815c1810
#> 58             bcEval 00007f57815ab2f0
#> 59            Rf_eval 00007f57815c0c60
#> 60       forcePromise 00007f57815c16d0
#> 61             getvar 00007f57815c1810
#> 62             bcEval 00007f57815ab2f0
#> 63            Rf_eval 00007f57815c0c60
#> 64       forcePromise 00007f57815c16d0
#> 65             getvar 00007f57815c1810
#> 66             bcEval 00007f57815ab2f0
#> 67            Rf_eval 00007f57815c0c60
#> 68       forcePromise 00007f57815c16d0
#> 69             getvar 00007f57815c1810
#> 70             bcEval 00007f57815ab2f0
#> 71            Rf_eval 00007f57815c0c60
#> 72      R_execClosure 00007f57815c28e0
#> 73    Rf_applyClosure 00007f57815c3800
#> 74             bcEval 00007f57815ab2f0
#> 75            Rf_eval 00007f57815c0c60
#> 76      R_execClosure 00007f57815c28e0
#> 77    Rf_applyClosure 00007f57815c3800
#> 78             bcEval 00007f57815ab2f0
#> 79            Rf_eval 00007f57815c0c60
#> 80      R_execClosure 00007f57815c28e0
#> 81    Rf_applyClosure 00007f57815c3800
#> 82             bcEval 00007f57815ab2f0
#> 83            Rf_eval 00007f57815c0c60
#> 84      R_execClosure 00007f57815c28e0
#> 85    Rf_applyClosure 00007f57815c3800
#> 86             bcEval 00007f57815ab2f0
#> 87            Rf_eval 00007f57815c0c60
#> 88      R_execClosure 00007f57815c28e0
#> 89    Rf_applyClosure 00007f57815c3800
#> 90             bcEval 00007f57815ab2f0
#> 91            Rf_eval 00007f57815c0c60
#> 92      R_execClosure 00007f57815c28e0
#> 93    Rf_applyClosure 00007f57815c3800
#> 94             bcEval 00007f57815ab2f0
#> 95            Rf_eval 00007f57815c0c60
#> 96       forcePromise 00007f57815c16d0
#> 97             getvar 00007f57815c1810
#> 98             bcEval 00007f57815ab2f0
#> 99            Rf_eval 00007f57815c0c60
#> 100     R_execClosure 00007f57815c28e0
#> 101   Rf_applyClosure 00007f57815c3800
#> 102            bcEval 00007f57815ab2f0
#> 103           Rf_eval 00007f57815c0c60
#> 104     R_execClosure 00007f57815c28e0
#> 105   Rf_applyClosure 00007f57815c3800
#> 106            bcEval 00007f57815ab2f0
#> 107           Rf_eval 00007f57815c0c60
#> 108     R_execClosure 00007f57815c28e0
#> 109   Rf_applyClosure 00007f57815c3800
#> 110            bcEval 00007f57815ab2f0
#> 111           Rf_eval 00007f57815c0c60
#> 112     R_execClosure 00007f57815c28e0
#> 113   Rf_applyClosure 00007f57815c3800
#> 114            bcEval 00007f57815ab2f0
#> 115           Rf_eval 00007f57815c0c60
#> 116     R_execClosure 00007f57815c28e0
#> 117   Rf_applyClosure 00007f57815c3800
#> 118            bcEval 00007f57815ab2f0
#> 119           Rf_eval 00007f57815c0c60
#> 120     R_execClosure 00007f57815c28e0
#> 121   Rf_applyClosure 00007f57815c3800
#> 122            bcEval 00007f57815ab2f0
#> 123           Rf_eval 00007f57815c0c60
#> 124     R_execClosure 00007f57815c28e0
#> 125   Rf_applyClosure 00007f57815c3800
#> 126            bcEval 00007f57815ab2f0
#> 127           Rf_eval 00007f57815c0c60
#> 128      forcePromise 00007f57815c16d0
#> 129            getvar 00007f57815c1810
#> 130            bcEval 00007f57815ab2f0
#> 131           Rf_eval 00007f57815c0c60
#> 132     R_execClosure 00007f57815c28e0
#> 133   Rf_applyClosure 00007f57815c3800
#> 134            bcEval 00007f57815ab2f0
#> 135           Rf_eval 00007f57815c0c60
#> 136     R_execClosure 00007f57815c28e0
#> 137   Rf_applyClosure 00007f57815c3800
#> 138    R_forceAndCall 00007f57815c61d0
#> 139         call_loop 00007f576fcf43b0
#> 140          map_impl 00007f576fcf4500
#> 141       R_doDotCall 00007f5781575010
#> 142            bcEval 00007f57815ab2f0
#> 143           Rf_eval 00007f57815c0c60
#> 144     R_execClosure 00007f57815c28e0
#> 145   Rf_applyClosure 00007f57815c3800
#> 146            bcEval 00007f57815ab2f0
#> 147           Rf_eval 00007f57815c0c60
#> 148     R_execClosure 00007f57815c28e0
#> 149   Rf_applyClosure 00007f57815c3800
#> 150            bcEval 00007f57815ab2f0
#> 151           Rf_eval 00007f57815c0c60
#> 152     R_execClosure 00007f57815c28e0
#> 153   Rf_applyClosure 00007f57815c3800
#> 154            bcEval 00007f57815ab2f0
#> 155           Rf_eval 00007f57815c0c60
#> 156     R_execClosure 00007f57815c28e0
#> 157   Rf_applyClosure 00007f57815c3800
#> 158            bcEval 00007f57815ab2f0
#> 159           Rf_eval 00007f57815c0c60
#> 160     R_execClosure 00007f57815c28e0
#> 161   Rf_applyClosure 00007f57815c3800
#> 162           Rf_eval 00007f57815c0c60
#> 163  Rf_ReplIteration 00007f57815f4900
#> 164     R_ReplConsole 00007f57815f4e80
#> 165     run_Rmainloop 00007f57815f4f70
#> 166              main 0000562713ebd820
#> 167 __libc_start_main 00007f57810a7b10
#> 168            _start 0000562713ebd850
#>                                                pathname
#> 1                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 2                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 3                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 4                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 5                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 6                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 7                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 8                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 9                        /opt/R/4.0.3/lib/R/lib/libR.so
#> 10                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 11  /home/runner/work/_temp/Library/winch/libs/winch.so
#> 12                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 13                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 14                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 15                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 16                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 17                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 18                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 19                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 20                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 21                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 22                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 23                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 24                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 25                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 26                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 27                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 28                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 29                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 30                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 31                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 32                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 33                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 34                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 35                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 36                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 37                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 38                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 39                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 40                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 41                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 42                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 43                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 44                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 45                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 46                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 47                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 48                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 49                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 50                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 51                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 52                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 53                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 54                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 55                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 56                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 57                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 58                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 59                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 60                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 61                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 62                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 63                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 64                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 65                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 66                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 67                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 68                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 69                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 70                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 71                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 72                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 73                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 74                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 75                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 76                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 77                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 78                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 79                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 80                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 81                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 82                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 83                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 84                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 85                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 86                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 87                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 88                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 89                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 90                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 91                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 92                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 93                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 94                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 95                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 96                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 97                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 98                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 99                       /opt/R/4.0.3/lib/R/lib/libR.so
#> 100                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 101                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 102                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 103                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 104                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 105                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 106                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 107                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 108                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 109                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 110                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 111                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 112                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 113                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 114                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 115                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 116                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 117                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 118                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 119                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 120                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 121                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 122                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 123                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 124                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 125                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 126                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 127                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 128                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 129                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 130                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 131                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 132                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 133                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 134                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 135                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 136                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 137                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 138                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 139 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 140 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 141                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 142                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 143                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 144                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 145                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 146                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 147                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 148                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 149                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 150                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 151                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 152                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 153                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 154                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 155                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 156                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 157                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 158                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 159                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 160                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 161                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 162                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 163                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 164                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 165                      /opt/R/4.0.3/lib/R/lib/libR.so
#> 166                       /opt/R/4.0.3/lib/R/bin/exec/R
#> 167                  /lib/x86_64-linux-gnu/libc-2.27.so
#> 168                       /opt/R/4.0.3/lib/R/bin/exec/R