Native stack trace

This function returns the native stack trace as a data frame. Each native stack frame corresponds to one row in the returned data frame. Deep function calls come first, the last row corresponds to the running process's entry point.

winch_trace_back()

Value

A data frame with the columns:

func: function name

ip: instruction pointer

pathname: path to shared library

Details

On Windows, call winch_init_library() to return function names for a specific package.

See also

Examples

winch_trace_back()
#>                   func               ip
#> 1             do_Rprof 00007f0eb3d108b0
#> 2              Rf_eval 00007f0eb3d2f8c0
#> 3              Rf_eval 00007f0eb3d31560
#> 4      Rf_applyClosure 00007f0eb3d32560
#> 5              Rf_eval 00007f0eb3d2f8c0
#> 6       R_forceAndCall 00007f0eb3d35780
#> 7             do_Rprof 00007f0eb3d108b0
#> 8              Rf_eval 00007f0eb3d2f8c0
#> 9              Rf_eval 00007f0eb3d31560
#> 10     Rf_applyClosure 00007f0eb3d32560
#> 11            do_Rprof 00007f0eb3d108b0
#> 12             Rf_eval 00007f0eb3d2f8c0
#> 13             Rf_eval 00007f0eb3d31560
#> 14     Rf_applyClosure 00007f0eb3d32560
#> 15            do_Rprof 00007f0eb3d108b0
#> 16             Rf_eval 00007f0eb3d2f8c0
#> 17             Rf_eval 00007f0eb3d30310
#> 18             Rf_eval 00007f0eb3d2f8c0
#> 19      R_forceAndCall 00007f0eb3d36530
#> 20         do_Rprofmem 00007f0eb3d740d0
#> 21            do_Rprof 00007f0eb3d108b0
#> 22             Rf_eval 00007f0eb3d2f8c0
#> 23             Rf_eval 00007f0eb3d31560
#> 24     Rf_applyClosure 00007f0eb3d32560
#> 25            do_Rprof 00007f0eb3d108b0
#> 26             Rf_eval 00007f0eb3d2f8c0
#> 27             Rf_eval 00007f0eb3d30310
#> 28             Rf_eval 00007f0eb3d30450
#> 29            do_Rprof 00007f0eb3d108b0
#> 30             Rf_eval 00007f0eb3d2f8c0
#> 31             Rf_eval 00007f0eb3d31560
#> 32     Rf_applyClosure 00007f0eb3d32560
#> 33            do_Rprof 00007f0eb3d108b0
#> 34             Rf_eval 00007f0eb3d2f8c0
#> 35             Rf_eval 00007f0eb3d30310
#> 36             Rf_eval 00007f0eb3d30450
#> 37            do_Rprof 00007f0eb3d108b0
#> 38             Rf_eval 00007f0eb3d2f8c0
#> 39             Rf_eval 00007f0eb3d30310
#> 40             Rf_eval 00007f0eb3d30450
#> 41            do_Rprof 00007f0eb3d108b0
#> 42             Rf_eval 00007f0eb3d2f8c0
#> 43             Rf_eval 00007f0eb3d30310
#> 44             Rf_eval 00007f0eb3d30450
#> 45            do_Rprof 00007f0eb3d108b0
#> 46             Rf_eval 00007f0eb3d2f8c0
#> 47             Rf_eval 00007f0eb3d30310
#> 48             Rf_eval 00007f0eb3d30450
#> 49            do_Rprof 00007f0eb3d108b0
#> 50             Rf_eval 00007f0eb3d2f8c0
#> 51             Rf_eval 00007f0eb3d30310
#> 52             Rf_eval 00007f0eb3d30450
#> 53            do_Rprof 00007f0eb3d108b0
#> 54             Rf_eval 00007f0eb3d2f8c0
#> 55             Rf_eval 00007f0eb3d30310
#> 56             Rf_eval 00007f0eb3d30450
#> 57            do_Rprof 00007f0eb3d108b0
#> 58             Rf_eval 00007f0eb3d2f8c0
#> 59             Rf_eval 00007f0eb3d31560
#> 60     Rf_applyClosure 00007f0eb3d32560
#> 61            do_Rprof 00007f0eb3d108b0
#> 62             Rf_eval 00007f0eb3d2f8c0
#> 63             Rf_eval 00007f0eb3d31560
#> 64     Rf_applyClosure 00007f0eb3d32560
#> 65            do_Rprof 00007f0eb3d108b0
#> 66             Rf_eval 00007f0eb3d2f8c0
#> 67             Rf_eval 00007f0eb3d31560
#> 68     Rf_applyClosure 00007f0eb3d32560
#> 69            do_Rprof 00007f0eb3d108b0
#> 70             Rf_eval 00007f0eb3d2f8c0
#> 71             Rf_eval 00007f0eb3d31560
#> 72     Rf_applyClosure 00007f0eb3d32560
#> 73            do_Rprof 00007f0eb3d108b0
#> 74             Rf_eval 00007f0eb3d2f8c0
#> 75             Rf_eval 00007f0eb3d31560
#> 76     Rf_applyClosure 00007f0eb3d32560
#> 77            do_Rprof 00007f0eb3d108b0
#> 78             Rf_eval 00007f0eb3d2f8c0
#> 79             Rf_eval 00007f0eb3d31560
#> 80     Rf_applyClosure 00007f0eb3d32560
#> 81            do_Rprof 00007f0eb3d108b0
#> 82             Rf_eval 00007f0eb3d2f8c0
#> 83             Rf_eval 00007f0eb3d30310
#> 84             Rf_eval 00007f0eb3d30450
#> 85            do_Rprof 00007f0eb3d108b0
#> 86             Rf_eval 00007f0eb3d2f8c0
#> 87             Rf_eval 00007f0eb3d31560
#> 88     Rf_applyClosure 00007f0eb3d32560
#> 89            do_Rprof 00007f0eb3d108b0
#> 90             Rf_eval 00007f0eb3d2f8c0
#> 91             Rf_eval 00007f0eb3d31560
#> 92     Rf_applyClosure 00007f0eb3d32560
#> 93            do_Rprof 00007f0eb3d108b0
#> 94             Rf_eval 00007f0eb3d2f8c0
#> 95             Rf_eval 00007f0eb3d31560
#> 96     Rf_applyClosure 00007f0eb3d32560
#> 97            do_Rprof 00007f0eb3d108b0
#> 98             Rf_eval 00007f0eb3d2f8c0
#> 99             Rf_eval 00007f0eb3d31560
#> 100    Rf_applyClosure 00007f0eb3d32560
#> 101           do_Rprof 00007f0eb3d108b0
#> 102            Rf_eval 00007f0eb3d2f8c0
#> 103            Rf_eval 00007f0eb3d31560
#> 104    Rf_applyClosure 00007f0eb3d32560
#> 105           do_Rprof 00007f0eb3d108b0
#> 106            Rf_eval 00007f0eb3d2f8c0
#> 107            Rf_eval 00007f0eb3d31560
#> 108    Rf_applyClosure 00007f0eb3d32560
#> 109           do_Rprof 00007f0eb3d108b0
#> 110            Rf_eval 00007f0eb3d2f8c0
#> 111            Rf_eval 00007f0eb3d31560
#> 112    Rf_applyClosure 00007f0eb3d32560
#> 113           do_Rprof 00007f0eb3d108b0
#> 114            Rf_eval 00007f0eb3d2f8c0
#> 115            Rf_eval 00007f0eb3d30310
#> 116            Rf_eval 00007f0eb3d30450
#> 117           do_Rprof 00007f0eb3d108b0
#> 118            Rf_eval 00007f0eb3d2f8c0
#> 119            Rf_eval 00007f0eb3d31560
#> 120    Rf_applyClosure 00007f0eb3d32560
#> 121           do_Rprof 00007f0eb3d108b0
#> 122            Rf_eval 00007f0eb3d2f8c0
#> 123            Rf_eval 00007f0eb3d31560
#> 124    Rf_applyClosure 00007f0eb3d32560
#> 125     R_forceAndCall 00007f0eb3d34ed0
#> 126          call_loop 00007f0eaaac4670
#> 127           map_impl 00007f0eaaac47d0
#> 128 Rf_NewFrameConfirm 00007f0eb3cd7b10
#> 129           do_Rprof 00007f0eb3d108b0
#> 130            Rf_eval 00007f0eb3d2f8c0
#> 131            Rf_eval 00007f0eb3d31560
#> 132    Rf_applyClosure 00007f0eb3d32560
#> 133           do_Rprof 00007f0eb3d108b0
#> 134            Rf_eval 00007f0eb3d2f8c0
#> 135            Rf_eval 00007f0eb3d31560
#> 136    Rf_applyClosure 00007f0eb3d32560
#> 137           do_Rprof 00007f0eb3d108b0
#> 138            Rf_eval 00007f0eb3d2f8c0
#> 139            Rf_eval 00007f0eb3d31560
#> 140    Rf_applyClosure 00007f0eb3d32560
#> 141           do_Rprof 00007f0eb3d108b0
#> 142            Rf_eval 00007f0eb3d2f8c0
#> 143            Rf_eval 00007f0eb3d31560
#> 144    Rf_applyClosure 00007f0eb3d32560
#> 145           do_Rprof 00007f0eb3d108b0
#> 146            Rf_eval 00007f0eb3d2f8c0
#> 147            Rf_eval 00007f0eb3d31560
#> 148    Rf_applyClosure 00007f0eb3d32560
#> 149           do_Rprof 00007f0eb3d108b0
#> 150            Rf_eval 00007f0eb3d2f8c0
#> 151            Rf_eval 00007f0eb3d31560
#> 152    Rf_applyClosure 00007f0eb3d32560
#> 153            Rf_eval 00007f0eb3d2f8c0
#> 154   Rf_ReplIteration 00007f0eb3d63c60
#> 155   Rf_ReplIteration 00007f0eb3d64170
#> 156      run_Rmainloop 00007f0eb3d64270
#> 157               main 00005625cee05080
#> 158  __libc_start_main 00007f0eb3a16fc0
#> 159             _start 00005625cee050b0
#>                                                pathname
#> 1                                /usr/lib/R/lib/libR.so
#> 2                                /usr/lib/R/lib/libR.so
#> 3                                /usr/lib/R/lib/libR.so
#> 4                                /usr/lib/R/lib/libR.so
#> 5                                /usr/lib/R/lib/libR.so
#> 6                                /usr/lib/R/lib/libR.so
#> 7                                /usr/lib/R/lib/libR.so
#> 8                                /usr/lib/R/lib/libR.so
#> 9                                /usr/lib/R/lib/libR.so
#> 10                               /usr/lib/R/lib/libR.so
#> 11                               /usr/lib/R/lib/libR.so
#> 12                               /usr/lib/R/lib/libR.so
#> 13                               /usr/lib/R/lib/libR.so
#> 14                               /usr/lib/R/lib/libR.so
#> 15                               /usr/lib/R/lib/libR.so
#> 16                               /usr/lib/R/lib/libR.so
#> 17                               /usr/lib/R/lib/libR.so
#> 18                               /usr/lib/R/lib/libR.so
#> 19                               /usr/lib/R/lib/libR.so
#> 20                               /usr/lib/R/lib/libR.so
#> 21                               /usr/lib/R/lib/libR.so
#> 22                               /usr/lib/R/lib/libR.so
#> 23                               /usr/lib/R/lib/libR.so
#> 24                               /usr/lib/R/lib/libR.so
#> 25                               /usr/lib/R/lib/libR.so
#> 26                               /usr/lib/R/lib/libR.so
#> 27                               /usr/lib/R/lib/libR.so
#> 28                               /usr/lib/R/lib/libR.so
#> 29                               /usr/lib/R/lib/libR.so
#> 30                               /usr/lib/R/lib/libR.so
#> 31                               /usr/lib/R/lib/libR.so
#> 32                               /usr/lib/R/lib/libR.so
#> 33                               /usr/lib/R/lib/libR.so
#> 34                               /usr/lib/R/lib/libR.so
#> 35                               /usr/lib/R/lib/libR.so
#> 36                               /usr/lib/R/lib/libR.so
#> 37                               /usr/lib/R/lib/libR.so
#> 38                               /usr/lib/R/lib/libR.so
#> 39                               /usr/lib/R/lib/libR.so
#> 40                               /usr/lib/R/lib/libR.so
#> 41                               /usr/lib/R/lib/libR.so
#> 42                               /usr/lib/R/lib/libR.so
#> 43                               /usr/lib/R/lib/libR.so
#> 44                               /usr/lib/R/lib/libR.so
#> 45                               /usr/lib/R/lib/libR.so
#> 46                               /usr/lib/R/lib/libR.so
#> 47                               /usr/lib/R/lib/libR.so
#> 48                               /usr/lib/R/lib/libR.so
#> 49                               /usr/lib/R/lib/libR.so
#> 50                               /usr/lib/R/lib/libR.so
#> 51                               /usr/lib/R/lib/libR.so
#> 52                               /usr/lib/R/lib/libR.so
#> 53                               /usr/lib/R/lib/libR.so
#> 54                               /usr/lib/R/lib/libR.so
#> 55                               /usr/lib/R/lib/libR.so
#> 56                               /usr/lib/R/lib/libR.so
#> 57                               /usr/lib/R/lib/libR.so
#> 58                               /usr/lib/R/lib/libR.so
#> 59                               /usr/lib/R/lib/libR.so
#> 60                               /usr/lib/R/lib/libR.so
#> 61                               /usr/lib/R/lib/libR.so
#> 62                               /usr/lib/R/lib/libR.so
#> 63                               /usr/lib/R/lib/libR.so
#> 64                               /usr/lib/R/lib/libR.so
#> 65                               /usr/lib/R/lib/libR.so
#> 66                               /usr/lib/R/lib/libR.so
#> 67                               /usr/lib/R/lib/libR.so
#> 68                               /usr/lib/R/lib/libR.so
#> 69                               /usr/lib/R/lib/libR.so
#> 70                               /usr/lib/R/lib/libR.so
#> 71                               /usr/lib/R/lib/libR.so
#> 72                               /usr/lib/R/lib/libR.so
#> 73                               /usr/lib/R/lib/libR.so
#> 74                               /usr/lib/R/lib/libR.so
#> 75                               /usr/lib/R/lib/libR.so
#> 76                               /usr/lib/R/lib/libR.so
#> 77                               /usr/lib/R/lib/libR.so
#> 78                               /usr/lib/R/lib/libR.so
#> 79                               /usr/lib/R/lib/libR.so
#> 80                               /usr/lib/R/lib/libR.so
#> 81                               /usr/lib/R/lib/libR.so
#> 82                               /usr/lib/R/lib/libR.so
#> 83                               /usr/lib/R/lib/libR.so
#> 84                               /usr/lib/R/lib/libR.so
#> 85                               /usr/lib/R/lib/libR.so
#> 86                               /usr/lib/R/lib/libR.so
#> 87                               /usr/lib/R/lib/libR.so
#> 88                               /usr/lib/R/lib/libR.so
#> 89                               /usr/lib/R/lib/libR.so
#> 90                               /usr/lib/R/lib/libR.so
#> 91                               /usr/lib/R/lib/libR.so
#> 92                               /usr/lib/R/lib/libR.so
#> 93                               /usr/lib/R/lib/libR.so
#> 94                               /usr/lib/R/lib/libR.so
#> 95                               /usr/lib/R/lib/libR.so
#> 96                               /usr/lib/R/lib/libR.so
#> 97                               /usr/lib/R/lib/libR.so
#> 98                               /usr/lib/R/lib/libR.so
#> 99                               /usr/lib/R/lib/libR.so
#> 100                              /usr/lib/R/lib/libR.so
#> 101                              /usr/lib/R/lib/libR.so
#> 102                              /usr/lib/R/lib/libR.so
#> 103                              /usr/lib/R/lib/libR.so
#> 104                              /usr/lib/R/lib/libR.so
#> 105                              /usr/lib/R/lib/libR.so
#> 106                              /usr/lib/R/lib/libR.so
#> 107                              /usr/lib/R/lib/libR.so
#> 108                              /usr/lib/R/lib/libR.so
#> 109                              /usr/lib/R/lib/libR.so
#> 110                              /usr/lib/R/lib/libR.so
#> 111                              /usr/lib/R/lib/libR.so
#> 112                              /usr/lib/R/lib/libR.so
#> 113                              /usr/lib/R/lib/libR.so
#> 114                              /usr/lib/R/lib/libR.so
#> 115                              /usr/lib/R/lib/libR.so
#> 116                              /usr/lib/R/lib/libR.so
#> 117                              /usr/lib/R/lib/libR.so
#> 118                              /usr/lib/R/lib/libR.so
#> 119                              /usr/lib/R/lib/libR.so
#> 120                              /usr/lib/R/lib/libR.so
#> 121                              /usr/lib/R/lib/libR.so
#> 122                              /usr/lib/R/lib/libR.so
#> 123                              /usr/lib/R/lib/libR.so
#> 124                              /usr/lib/R/lib/libR.so
#> 125                              /usr/lib/R/lib/libR.so
#> 126 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 127 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 128                              /usr/lib/R/lib/libR.so
#> 129                              /usr/lib/R/lib/libR.so
#> 130                              /usr/lib/R/lib/libR.so
#> 131                              /usr/lib/R/lib/libR.so
#> 132                              /usr/lib/R/lib/libR.so
#> 133                              /usr/lib/R/lib/libR.so
#> 134                              /usr/lib/R/lib/libR.so
#> 135                              /usr/lib/R/lib/libR.so
#> 136                              /usr/lib/R/lib/libR.so
#> 137                              /usr/lib/R/lib/libR.so
#> 138                              /usr/lib/R/lib/libR.so
#> 139                              /usr/lib/R/lib/libR.so
#> 140                              /usr/lib/R/lib/libR.so
#> 141                              /usr/lib/R/lib/libR.so
#> 142                              /usr/lib/R/lib/libR.so
#> 143                              /usr/lib/R/lib/libR.so
#> 144                              /usr/lib/R/lib/libR.so
#> 145                              /usr/lib/R/lib/libR.so
#> 146                              /usr/lib/R/lib/libR.so
#> 147                              /usr/lib/R/lib/libR.so
#> 148                              /usr/lib/R/lib/libR.so
#> 149                              /usr/lib/R/lib/libR.so
#> 150                              /usr/lib/R/lib/libR.so
#> 151                              /usr/lib/R/lib/libR.so
#> 152                              /usr/lib/R/lib/libR.so
#> 153                              /usr/lib/R/lib/libR.so
#> 154                              /usr/lib/R/lib/libR.so
#> 155                              /usr/lib/R/lib/libR.so
#> 156                              /usr/lib/R/lib/libR.so
#> 157                               /usr/lib/R/bin/exec/R
#> 158              /usr/lib/x86_64-linux-gnu/libc-2.31.so
#> 159                               /usr/lib/R/bin/exec/R

foo <- function() {
  winch_call(bar)
}

bar <- function() {
  winch_trace_back()
}

foo()
#>                   func               ip
#> 1             do_Rprof 00007f0eb3d108b0
#> 2              Rf_eval 00007f0eb3d2f8c0
#> 3              Rf_eval 00007f0eb3d31560
#> 4      Rf_applyClosure 00007f0eb3d32560
#> 5              Rf_eval 00007f0eb3d2f8c0
#> 6         R_execMethod 00007f0eb3d33350
#> 7              Rf_eval 00007f0eb3d2f8c0
#> 8              Rf_eval 00007f0eb3d31560
#> 9      Rf_applyClosure 00007f0eb3d32560
#> 10             Rf_eval 00007f0eb3d2f8c0
#> 11          winch_call 00007f0ea87c7720
#> 12  Rf_NewFrameConfirm 00007f0eb3cd7b10
#> 13            do_Rprof 00007f0eb3d108b0
#> 14             Rf_eval 00007f0eb3d2f8c0
#> 15             Rf_eval 00007f0eb3d31560
#> 16     Rf_applyClosure 00007f0eb3d32560
#> 17             Rf_eval 00007f0eb3d2f8c0
#> 18        R_execMethod 00007f0eb3d33350
#> 19             Rf_eval 00007f0eb3d2f8c0
#> 20             Rf_eval 00007f0eb3d31560
#> 21     Rf_applyClosure 00007f0eb3d32560
#> 22             Rf_eval 00007f0eb3d2f8c0
#> 23      R_forceAndCall 00007f0eb3d35780
#> 24            do_Rprof 00007f0eb3d108b0
#> 25             Rf_eval 00007f0eb3d2f8c0
#> 26             Rf_eval 00007f0eb3d31560
#> 27     Rf_applyClosure 00007f0eb3d32560
#> 28            do_Rprof 00007f0eb3d108b0
#> 29             Rf_eval 00007f0eb3d2f8c0
#> 30             Rf_eval 00007f0eb3d31560
#> 31     Rf_applyClosure 00007f0eb3d32560
#> 32            do_Rprof 00007f0eb3d108b0
#> 33             Rf_eval 00007f0eb3d2f8c0
#> 34             Rf_eval 00007f0eb3d30310
#> 35             Rf_eval 00007f0eb3d2f8c0
#> 36      R_forceAndCall 00007f0eb3d36530
#> 37         do_Rprofmem 00007f0eb3d740d0
#> 38            do_Rprof 00007f0eb3d108b0
#> 39             Rf_eval 00007f0eb3d2f8c0
#> 40             Rf_eval 00007f0eb3d31560
#> 41     Rf_applyClosure 00007f0eb3d32560
#> 42            do_Rprof 00007f0eb3d108b0
#> 43             Rf_eval 00007f0eb3d2f8c0
#> 44             Rf_eval 00007f0eb3d30310
#> 45             Rf_eval 00007f0eb3d30450
#> 46            do_Rprof 00007f0eb3d108b0
#> 47             Rf_eval 00007f0eb3d2f8c0
#> 48             Rf_eval 00007f0eb3d31560
#> 49     Rf_applyClosure 00007f0eb3d32560
#> 50            do_Rprof 00007f0eb3d108b0
#> 51             Rf_eval 00007f0eb3d2f8c0
#> 52             Rf_eval 00007f0eb3d30310
#> 53             Rf_eval 00007f0eb3d30450
#> 54            do_Rprof 00007f0eb3d108b0
#> 55             Rf_eval 00007f0eb3d2f8c0
#> 56             Rf_eval 00007f0eb3d30310
#> 57             Rf_eval 00007f0eb3d30450
#> 58            do_Rprof 00007f0eb3d108b0
#> 59             Rf_eval 00007f0eb3d2f8c0
#> 60             Rf_eval 00007f0eb3d30310
#> 61             Rf_eval 00007f0eb3d30450
#> 62            do_Rprof 00007f0eb3d108b0
#> 63             Rf_eval 00007f0eb3d2f8c0
#> 64             Rf_eval 00007f0eb3d30310
#> 65             Rf_eval 00007f0eb3d30450
#> 66            do_Rprof 00007f0eb3d108b0
#> 67             Rf_eval 00007f0eb3d2f8c0
#> 68             Rf_eval 00007f0eb3d30310
#> 69             Rf_eval 00007f0eb3d30450
#> 70            do_Rprof 00007f0eb3d108b0
#> 71             Rf_eval 00007f0eb3d2f8c0
#> 72             Rf_eval 00007f0eb3d30310
#> 73             Rf_eval 00007f0eb3d30450
#> 74            do_Rprof 00007f0eb3d108b0
#> 75             Rf_eval 00007f0eb3d2f8c0
#> 76             Rf_eval 00007f0eb3d31560
#> 77     Rf_applyClosure 00007f0eb3d32560
#> 78            do_Rprof 00007f0eb3d108b0
#> 79             Rf_eval 00007f0eb3d2f8c0
#> 80             Rf_eval 00007f0eb3d31560
#> 81     Rf_applyClosure 00007f0eb3d32560
#> 82            do_Rprof 00007f0eb3d108b0
#> 83             Rf_eval 00007f0eb3d2f8c0
#> 84             Rf_eval 00007f0eb3d31560
#> 85     Rf_applyClosure 00007f0eb3d32560
#> 86            do_Rprof 00007f0eb3d108b0
#> 87             Rf_eval 00007f0eb3d2f8c0
#> 88             Rf_eval 00007f0eb3d31560
#> 89     Rf_applyClosure 00007f0eb3d32560
#> 90            do_Rprof 00007f0eb3d108b0
#> 91             Rf_eval 00007f0eb3d2f8c0
#> 92             Rf_eval 00007f0eb3d31560
#> 93     Rf_applyClosure 00007f0eb3d32560
#> 94            do_Rprof 00007f0eb3d108b0
#> 95             Rf_eval 00007f0eb3d2f8c0
#> 96             Rf_eval 00007f0eb3d31560
#> 97     Rf_applyClosure 00007f0eb3d32560
#> 98            do_Rprof 00007f0eb3d108b0
#> 99             Rf_eval 00007f0eb3d2f8c0
#> 100            Rf_eval 00007f0eb3d30310
#> 101            Rf_eval 00007f0eb3d30450
#> 102           do_Rprof 00007f0eb3d108b0
#> 103            Rf_eval 00007f0eb3d2f8c0
#> 104            Rf_eval 00007f0eb3d31560
#> 105    Rf_applyClosure 00007f0eb3d32560
#> 106           do_Rprof 00007f0eb3d108b0
#> 107            Rf_eval 00007f0eb3d2f8c0
#> 108            Rf_eval 00007f0eb3d31560
#> 109    Rf_applyClosure 00007f0eb3d32560
#> 110           do_Rprof 00007f0eb3d108b0
#> 111            Rf_eval 00007f0eb3d2f8c0
#> 112            Rf_eval 00007f0eb3d31560
#> 113    Rf_applyClosure 00007f0eb3d32560
#> 114           do_Rprof 00007f0eb3d108b0
#> 115            Rf_eval 00007f0eb3d2f8c0
#> 116            Rf_eval 00007f0eb3d31560
#> 117    Rf_applyClosure 00007f0eb3d32560
#> 118           do_Rprof 00007f0eb3d108b0
#> 119            Rf_eval 00007f0eb3d2f8c0
#> 120            Rf_eval 00007f0eb3d31560
#> 121    Rf_applyClosure 00007f0eb3d32560
#> 122           do_Rprof 00007f0eb3d108b0
#> 123            Rf_eval 00007f0eb3d2f8c0
#> 124            Rf_eval 00007f0eb3d31560
#> 125    Rf_applyClosure 00007f0eb3d32560
#> 126           do_Rprof 00007f0eb3d108b0
#> 127            Rf_eval 00007f0eb3d2f8c0
#> 128            Rf_eval 00007f0eb3d31560
#> 129    Rf_applyClosure 00007f0eb3d32560
#> 130           do_Rprof 00007f0eb3d108b0
#> 131            Rf_eval 00007f0eb3d2f8c0
#> 132            Rf_eval 00007f0eb3d30310
#> 133            Rf_eval 00007f0eb3d30450
#> 134           do_Rprof 00007f0eb3d108b0
#> 135            Rf_eval 00007f0eb3d2f8c0
#> 136            Rf_eval 00007f0eb3d31560
#> 137    Rf_applyClosure 00007f0eb3d32560
#> 138           do_Rprof 00007f0eb3d108b0
#> 139            Rf_eval 00007f0eb3d2f8c0
#> 140            Rf_eval 00007f0eb3d31560
#> 141    Rf_applyClosure 00007f0eb3d32560
#> 142     R_forceAndCall 00007f0eb3d34ed0
#> 143          call_loop 00007f0eaaac4670
#> 144           map_impl 00007f0eaaac47d0
#> 145 Rf_NewFrameConfirm 00007f0eb3cd7b10
#> 146           do_Rprof 00007f0eb3d108b0
#> 147            Rf_eval 00007f0eb3d2f8c0
#> 148            Rf_eval 00007f0eb3d31560
#> 149    Rf_applyClosure 00007f0eb3d32560
#> 150           do_Rprof 00007f0eb3d108b0
#> 151            Rf_eval 00007f0eb3d2f8c0
#> 152            Rf_eval 00007f0eb3d31560
#> 153    Rf_applyClosure 00007f0eb3d32560
#> 154           do_Rprof 00007f0eb3d108b0
#> 155            Rf_eval 00007f0eb3d2f8c0
#> 156            Rf_eval 00007f0eb3d31560
#> 157    Rf_applyClosure 00007f0eb3d32560
#> 158           do_Rprof 00007f0eb3d108b0
#> 159            Rf_eval 00007f0eb3d2f8c0
#> 160            Rf_eval 00007f0eb3d31560
#> 161    Rf_applyClosure 00007f0eb3d32560
#> 162           do_Rprof 00007f0eb3d108b0
#> 163            Rf_eval 00007f0eb3d2f8c0
#> 164            Rf_eval 00007f0eb3d31560
#> 165    Rf_applyClosure 00007f0eb3d32560
#> 166           do_Rprof 00007f0eb3d108b0
#> 167            Rf_eval 00007f0eb3d2f8c0
#> 168            Rf_eval 00007f0eb3d31560
#> 169    Rf_applyClosure 00007f0eb3d32560
#> 170            Rf_eval 00007f0eb3d2f8c0
#> 171   Rf_ReplIteration 00007f0eb3d63c60
#> 172   Rf_ReplIteration 00007f0eb3d64170
#> 173      run_Rmainloop 00007f0eb3d64270
#> 174               main 00005625cee05080
#> 175  __libc_start_main 00007f0eb3a16fc0
#> 176             _start 00005625cee050b0
#>                                                pathname
#> 1                                /usr/lib/R/lib/libR.so
#> 2                                /usr/lib/R/lib/libR.so
#> 3                                /usr/lib/R/lib/libR.so
#> 4                                /usr/lib/R/lib/libR.so
#> 5                                /usr/lib/R/lib/libR.so
#> 6                                /usr/lib/R/lib/libR.so
#> 7                                /usr/lib/R/lib/libR.so
#> 8                                /usr/lib/R/lib/libR.so
#> 9                                /usr/lib/R/lib/libR.so
#> 10                               /usr/lib/R/lib/libR.so
#> 11  /home/runner/work/_temp/Library/winch/libs/winch.so
#> 12                               /usr/lib/R/lib/libR.so
#> 13                               /usr/lib/R/lib/libR.so
#> 14                               /usr/lib/R/lib/libR.so
#> 15                               /usr/lib/R/lib/libR.so
#> 16                               /usr/lib/R/lib/libR.so
#> 17                               /usr/lib/R/lib/libR.so
#> 18                               /usr/lib/R/lib/libR.so
#> 19                               /usr/lib/R/lib/libR.so
#> 20                               /usr/lib/R/lib/libR.so
#> 21                               /usr/lib/R/lib/libR.so
#> 22                               /usr/lib/R/lib/libR.so
#> 23                               /usr/lib/R/lib/libR.so
#> 24                               /usr/lib/R/lib/libR.so
#> 25                               /usr/lib/R/lib/libR.so
#> 26                               /usr/lib/R/lib/libR.so
#> 27                               /usr/lib/R/lib/libR.so
#> 28                               /usr/lib/R/lib/libR.so
#> 29                               /usr/lib/R/lib/libR.so
#> 30                               /usr/lib/R/lib/libR.so
#> 31                               /usr/lib/R/lib/libR.so
#> 32                               /usr/lib/R/lib/libR.so
#> 33                               /usr/lib/R/lib/libR.so
#> 34                               /usr/lib/R/lib/libR.so
#> 35                               /usr/lib/R/lib/libR.so
#> 36                               /usr/lib/R/lib/libR.so
#> 37                               /usr/lib/R/lib/libR.so
#> 38                               /usr/lib/R/lib/libR.so
#> 39                               /usr/lib/R/lib/libR.so
#> 40                               /usr/lib/R/lib/libR.so
#> 41                               /usr/lib/R/lib/libR.so
#> 42                               /usr/lib/R/lib/libR.so
#> 43                               /usr/lib/R/lib/libR.so
#> 44                               /usr/lib/R/lib/libR.so
#> 45                               /usr/lib/R/lib/libR.so
#> 46                               /usr/lib/R/lib/libR.so
#> 47                               /usr/lib/R/lib/libR.so
#> 48                               /usr/lib/R/lib/libR.so
#> 49                               /usr/lib/R/lib/libR.so
#> 50                               /usr/lib/R/lib/libR.so
#> 51                               /usr/lib/R/lib/libR.so
#> 52                               /usr/lib/R/lib/libR.so
#> 53                               /usr/lib/R/lib/libR.so
#> 54                               /usr/lib/R/lib/libR.so
#> 55                               /usr/lib/R/lib/libR.so
#> 56                               /usr/lib/R/lib/libR.so
#> 57                               /usr/lib/R/lib/libR.so
#> 58                               /usr/lib/R/lib/libR.so
#> 59                               /usr/lib/R/lib/libR.so
#> 60                               /usr/lib/R/lib/libR.so
#> 61                               /usr/lib/R/lib/libR.so
#> 62                               /usr/lib/R/lib/libR.so
#> 63                               /usr/lib/R/lib/libR.so
#> 64                               /usr/lib/R/lib/libR.so
#> 65                               /usr/lib/R/lib/libR.so
#> 66                               /usr/lib/R/lib/libR.so
#> 67                               /usr/lib/R/lib/libR.so
#> 68                               /usr/lib/R/lib/libR.so
#> 69                               /usr/lib/R/lib/libR.so
#> 70                               /usr/lib/R/lib/libR.so
#> 71                               /usr/lib/R/lib/libR.so
#> 72                               /usr/lib/R/lib/libR.so
#> 73                               /usr/lib/R/lib/libR.so
#> 74                               /usr/lib/R/lib/libR.so
#> 75                               /usr/lib/R/lib/libR.so
#> 76                               /usr/lib/R/lib/libR.so
#> 77                               /usr/lib/R/lib/libR.so
#> 78                               /usr/lib/R/lib/libR.so
#> 79                               /usr/lib/R/lib/libR.so
#> 80                               /usr/lib/R/lib/libR.so
#> 81                               /usr/lib/R/lib/libR.so
#> 82                               /usr/lib/R/lib/libR.so
#> 83                               /usr/lib/R/lib/libR.so
#> 84                               /usr/lib/R/lib/libR.so
#> 85                               /usr/lib/R/lib/libR.so
#> 86                               /usr/lib/R/lib/libR.so
#> 87                               /usr/lib/R/lib/libR.so
#> 88                               /usr/lib/R/lib/libR.so
#> 89                               /usr/lib/R/lib/libR.so
#> 90                               /usr/lib/R/lib/libR.so
#> 91                               /usr/lib/R/lib/libR.so
#> 92                               /usr/lib/R/lib/libR.so
#> 93                               /usr/lib/R/lib/libR.so
#> 94                               /usr/lib/R/lib/libR.so
#> 95                               /usr/lib/R/lib/libR.so
#> 96                               /usr/lib/R/lib/libR.so
#> 97                               /usr/lib/R/lib/libR.so
#> 98                               /usr/lib/R/lib/libR.so
#> 99                               /usr/lib/R/lib/libR.so
#> 100                              /usr/lib/R/lib/libR.so
#> 101                              /usr/lib/R/lib/libR.so
#> 102                              /usr/lib/R/lib/libR.so
#> 103                              /usr/lib/R/lib/libR.so
#> 104                              /usr/lib/R/lib/libR.so
#> 105                              /usr/lib/R/lib/libR.so
#> 106                              /usr/lib/R/lib/libR.so
#> 107                              /usr/lib/R/lib/libR.so
#> 108                              /usr/lib/R/lib/libR.so
#> 109                              /usr/lib/R/lib/libR.so
#> 110                              /usr/lib/R/lib/libR.so
#> 111                              /usr/lib/R/lib/libR.so
#> 112                              /usr/lib/R/lib/libR.so
#> 113                              /usr/lib/R/lib/libR.so
#> 114                              /usr/lib/R/lib/libR.so
#> 115                              /usr/lib/R/lib/libR.so
#> 116                              /usr/lib/R/lib/libR.so
#> 117                              /usr/lib/R/lib/libR.so
#> 118                              /usr/lib/R/lib/libR.so
#> 119                              /usr/lib/R/lib/libR.so
#> 120                              /usr/lib/R/lib/libR.so
#> 121                              /usr/lib/R/lib/libR.so
#> 122                              /usr/lib/R/lib/libR.so
#> 123                              /usr/lib/R/lib/libR.so
#> 124                              /usr/lib/R/lib/libR.so
#> 125                              /usr/lib/R/lib/libR.so
#> 126                              /usr/lib/R/lib/libR.so
#> 127                              /usr/lib/R/lib/libR.so
#> 128                              /usr/lib/R/lib/libR.so
#> 129                              /usr/lib/R/lib/libR.so
#> 130                              /usr/lib/R/lib/libR.so
#> 131                              /usr/lib/R/lib/libR.so
#> 132                              /usr/lib/R/lib/libR.so
#> 133                              /usr/lib/R/lib/libR.so
#> 134                              /usr/lib/R/lib/libR.so
#> 135                              /usr/lib/R/lib/libR.so
#> 136                              /usr/lib/R/lib/libR.so
#> 137                              /usr/lib/R/lib/libR.so
#> 138                              /usr/lib/R/lib/libR.so
#> 139                              /usr/lib/R/lib/libR.so
#> 140                              /usr/lib/R/lib/libR.so
#> 141                              /usr/lib/R/lib/libR.so
#> 142                              /usr/lib/R/lib/libR.so
#> 143 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 144 /home/runner/work/_temp/Library/purrr/libs/purrr.so
#> 145                              /usr/lib/R/lib/libR.so
#> 146                              /usr/lib/R/lib/libR.so
#> 147                              /usr/lib/R/lib/libR.so
#> 148                              /usr/lib/R/lib/libR.so
#> 149                              /usr/lib/R/lib/libR.so
#> 150                              /usr/lib/R/lib/libR.so
#> 151                              /usr/lib/R/lib/libR.so
#> 152                              /usr/lib/R/lib/libR.so
#> 153                              /usr/lib/R/lib/libR.so
#> 154                              /usr/lib/R/lib/libR.so
#> 155                              /usr/lib/R/lib/libR.so
#> 156                              /usr/lib/R/lib/libR.so
#> 157                              /usr/lib/R/lib/libR.so
#> 158                              /usr/lib/R/lib/libR.so
#> 159                              /usr/lib/R/lib/libR.so
#> 160                              /usr/lib/R/lib/libR.so
#> 161                              /usr/lib/R/lib/libR.so
#> 162                              /usr/lib/R/lib/libR.so
#> 163                              /usr/lib/R/lib/libR.so
#> 164                              /usr/lib/R/lib/libR.so
#> 165                              /usr/lib/R/lib/libR.so
#> 166                              /usr/lib/R/lib/libR.so
#> 167                              /usr/lib/R/lib/libR.so
#> 168                              /usr/lib/R/lib/libR.so
#> 169                              /usr/lib/R/lib/libR.so
#> 170                              /usr/lib/R/lib/libR.so
#> 171                              /usr/lib/R/lib/libR.so
#> 172                              /usr/lib/R/lib/libR.so
#> 173                              /usr/lib/R/lib/libR.so
#> 174                               /usr/lib/R/bin/exec/R
#> 175              /usr/lib/x86_64-linux-gnu/libc-2.31.so
#> 176                               /usr/lib/R/bin/exec/R